==> scripts/1.MAY22_process.Rmd <==
---
title: "1.MAY22_process"
author: "abbey yatsko"
date: "2023-09-12"
output: html_document
---

libraries
```{r workspace prep}
library(data.table)
library(ggplot2)
library(ggpubr)
library(sqldf)
library(dplyr)
library(tidyverse)
library(stringr)
library(plyr)
library(gridExtra)
```

May 2022 data load and prep (manually fixing dates due to internal time clock error)
```{r data prep}
# all LGR files loading in
dataTMI <- list.files(path ="/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/LGR_files/may22LGR", pattern='f00', full.names= T) %>%
  setNames(nm = .) %>% 
  lapply(read.csv,skip=1) %>%
  bind_rows(.id = "file") %>%
  mutate(file = str_extract(file, "f00\\d+"))

# pgp error message occurs at the end of some LGR files - need to remove these lines 
data_clean <- subset(dataTMI, X.CH4._ppm!="NA")

# read in metadata to alter dates for each f#### file so that the fDOY makes sense 
meta_TMI <- read.csv("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/LGR_metadata/TMI_flux_metadata_may22.csv")
file <- unique(meta_TMI$file)
file_years <- as.data.frame(file)

# create repeating months 1-12 to make fDOY numbers have a smaller range 
file_years$month_alt <- rep(01:12, length.out=67)

# create new years so that each date is unique and not repeated for the fDOY
file_years$year_alt <- ""
file_years$year_alt[1:12] <- 2002
file_years$year_alt[13:24] <- 2003
file_years$year_alt[25:36] <- 2004
file_years$year_alt[37:48] <- 2005
file_years$year_alt[49:60] <- 2006
file_years$year_alt[61:67] <- 2007

# merge in alternative year and month to flux df
data_clean <- merge(data_clean, file_years, by = c("file")) 

# re-order month_alt year_alt 
data_clean <- data_clean %>% relocate(year_alt, .before = X.CH4._ppm)
data_clean <- data_clean %>% relocate(month_alt, .before = year_alt)

# create alternative times - replace original year (always 2002) with the created alternative year 
data_clean$Time <- gsub("2002", "", data_clean$Time)
data_clean$time_altyr <- with(data_clean, paste(substr(Time, 1, 8), year_alt, substr(Time, 9, nchar(Time)), sep = ""))

# remove first two characters from time_altmo column strings (for replacing month)
data_clean$time_altmo <- substring(data_clean$time_altyr , 5)

# paste in month_alt 
data_clean$time_alt <- with(data_clean, paste(substr(time_altmo, 1, 0), month_alt,
  substr(time_altmo, 1, nchar(time_altmo)), substr(time_altmo, 1, nchar(time_altmo)), sep = ""))

# move up time_alt to compare with original Time
data_clean <- data_clean %>% relocate(time_alt, .before = month_alt)

# select for useful columns 
data_clean <- data_clean[c("file", "time_alt", "X.CH4._ppm", "X.CH4._ppm_sd", "X.H2O._ppm", "X.H2O._ppm_sd", "X.CO2._ppm", "X.CO2._ppm_sd", "X.CH4.d_ppm", "X.CH4.d_ppm_sd", "X.CO2.d_ppm", "X.CO2.d_ppm_sd", "GasP_torr", "GasP_torr_sd", "GasT_C", "GasT_C_sd", "AmbT_C", "AmbT_C_sd", "RD0_us", "RD0_us_sd", "RD1_us", "RD1_us_sd", "Fit_Flag", "MIU_VALVE", "MIU_DESC")]

# rename "time_alt" to "Time"
data_clean <- dplyr::rename(data_clean, c("Time" = "time_alt"))

# format time 
date_time <- strptime(data_clean[,2],format='%m/%d/%Y %H:%M:%S')

# add year,month,day,JD,hour,min,sec columns to dataframe
Year <- as.numeric(format(date_time,'%Y'))
Month <- as.numeric(format(date_time,'%m'))
Day <- as.numeric(format(date_time,'%d'))
fDOY <- as.numeric(julian(date_time,'2002-01-01'))  #Change for year
Hour <- as.numeric(format(date_time,'%k'))
Min <- as.numeric(format(date_time,'%M'))
Sec <- as.numeric(format(date_time,'%S'))
data_clean <- cbind(date_time,Year,Month,Day,fDOY,Hour,Min,Sec,data_clean)

# save LGR data as data.table
data_clean <- data.table(data_clean)
```

Preliminary visualization for entire df (check and make sure fluxes show up)
```{r}
# for co2
co2 <- ggplot(data_clean, aes(date_time, X.CO2.d_ppm)) +
  geom_point() +
  theme(axis.text.x = element_text(angle = 90, hjust = 1)) +
  ggtitle("CO2 flux")

# for ch4 
ch4 <- ggplot(data_clean, aes(date_time, X.CH4.d_ppm)) +
  geom_point() +
  theme(axis.text.x = element_text(angle = 90, hjust = 1)) +
  ggtitle("CH4 flux")

# looks like there are the right number of lines that represent the 67 flux files 
```

Merge with metadata
```{r}
# merge in year_alt and month_alt to metadata
meta_TMI <- merge(meta_TMI, file_years, by = c("file")) 

# add in column for measurement_day_alt
meta_TMI$measurement_day_alt <- "/01/"

# paste in year_alt to the end of measurement_day_alt
meta_TMI$measurement_day_altyr <- with(meta_TMI, paste(substr(measurement_day_alt, 1, 8), year_alt, substr(measurement_day_alt, 9, nchar(measurement_day_alt)), sep = ""))

# paste in month_alt to beginning of  measurement_day_alt
meta_TMI$measurement_day_alt <- with(meta_TMI, paste(substr(measurement_day_altyr, 1, 0), month_alt, substr(measurement_day_altyr, 1, nchar(measurement_day_altyr)), sep = ""))

# reposition measurement_day_alt in df 
meta_TMI <- meta_TMI %>% relocate(measurement_day_alt, .before = measurement_day_actual)
meta_TMI <- meta_TMI %>% relocate(year_alt, .before = measurement_day_actual)
meta_TMI <- meta_TMI %>% relocate(month_alt, .before = year_alt)

# format measurement_day_alt
as.Date(meta_TMI$measurement_day_alt, '%m/%d/%Y')
```

Add start and end time for flux measurements 
```{r}
## Pull out date and time data
date_time_start <- strptime(paste(meta_TMI$measurement_day_alt,meta_TMI$flux_start),'%m/%d/%Y %H:%M:%S')

## Add 30 seconds to set buffered start time
date_time_start <- date_time_start+30

## Add 1.5 min (90 sec) to determine end time
date_time_end <- date_time_start+90

## Add fDOY columns for start and endtime to dataframe
fDOY_start <- as.numeric(julian(date_time_start,'2002-01-01'))  #Change for year
fDOY_end <- as.numeric(julian(date_time_end,'2002-01-01'))  #Change for year
meta_TMI <- cbind(fDOY_start,fDOY_end,meta_TMI)

# calculate average respiration temperature - add column
meta_TMI$avg_respT <- rowMeans(meta_TMI[ , c(13:16)], na.rm = TRUE)

# change sample to factor level variables 
meta_TMI$sample <- as.factor(meta_TMI$sample) 

# rename co2 and ch4 columns because naming format including '.' messes up the sql code 
data_clean <- data_clean %>%
  dplyr::rename(CH4 = X.CH4._ppm,
         CO2 = X.CO2._ppm, 
         CH4_dry = X.CH4.d_ppm,
         CO2_dry = X.CO2.d_ppm)
```

Merge LGR files with metadata 
```{r}
# complete merge for respiration data and metadata (may)
data_merge <- sqldf("select data_clean.fDOY, data_clean.CO2_dry, data_clean.CH4_dry, meta_TMI.sample, meta_TMI.avg_respT, meta_TMI.flux_source, meta_TMI.position, meta_TMI.tag, meta_TMI.directon,meta_TMI.chamber, meta_TMI.chamber_SA_cm2, meta_TMI.soil_moisture, meta_TMI.fDOY_start from data_clean LEFT JOIN meta_TMI ON (data_clean.fDOY BETWEEN meta_TMI.fDOY_start AND meta_TMI.fDOY_end)")

# filter out rows that  have no value for 'sample' - this means that no mound was being recorded
data_merge <- data_merge[complete.cases(data_merge[,c("sample")]),]

# check to make sure all of the mounds are there (67 total)
summary(unique(data_merge$sample))
```

Visually check clipped bits of data 
```{r}
# co2 (xlim=c(0.4247222, 0.4332870) corresponds to only MD1 first file)
par(mfrow=c(2,1),mar=c(4,4,1,1))
with(data_clean,plot(fDOY,CO2_dry,ylim=c(400,1300), xlim=c(0.4247222, 0.4332870)))
with(data_merge,plot(fDOY,CO2_dry,ylim=c(400,1300), xlim=c(0.4247222, 0.4332870)))

# ch4 
par(mfrow=c(2,1),mar=c(4,4,1,1))
with(data_clean,plot(fDOY,CH4_dry,ylim=c(1, 10), xlim=c(0.4247222, 0.4332870)))
with(data_merge,plot(fDOY,CH4_dry,ylim=c(1,10), xlim=c(0.4247222, 0.4332870)))
```

Convert ppm to moles for CH4 and CO2
```{r}
# use known volume of chamber (in m3) to convert ppm of gas to liters of gas
# small, h = 5.7 cm, d = 7.8 cm
sm <- pi*((7.8/2)^2)*5.7

# medium, h = 10.8 cm, d = 10.3 cm
med <- pi*((10.3/2)^2)*10.8

# large, h = 16.1 cm, d = 15.3 cm
lg <- pi*((15.3/2)^2)*16.1

# mini, v = 119 ml
mini <- 119

volume_cm3 <- c(sm, med, lg, mini)
chamber <- c("sm", "med", "lg", "mini")
chamber_vols <- data.frame(volume_cm3, chamber)

# chamber volumes are in cm3 - convert to m3 for the next step 
chamber_vols$volume_m3 <- chamber_vols$volume_cm3*0.000001

data_merge <- merge(data_merge, chamber_vols, by = c("chamber")) 
data_merge <- data_merge %>% relocate(chamber, .after = chamber_SA_cm2)

# also convert SA chamber to be in m2 (currently in cm2)
data_merge$chamber_SA_m2 <- data_merge$chamber_SA_cm2 * 0.0001
data_merge <- data_merge %>% relocate(chamber_SA_m2, .after = chamber_SA_cm2)

# CH4
data_merge$CH4_dry_L=
  # parts CH4 per million parts air * volume of air in chamber (m3) * 1000 L per m3
  (data_merge$CH4_dry/1000000) * data_merge$volume_m3 * 1000 

# CO2
data_merge$CO2_dry_L=
  # parts CO2 per million parts air * volume of air in chamber (m3) * 1000 L per m3
  (data_merge$CO2_dry/1000000) * data_merge$volume_m3 * 1000 

# Use ideal gas law to calculate umol of CH4 or mmol of CO2
# CH4
data_merge$CH4_dry_umol=
  # (atm pressure * L CH4) / (R in L*atm/?K*mol * ?K temp) * 10^6 umol/mol
  ((1*data_merge$CH4_dry_L)/(0.08206*(data_merge$avg_respT+273)))*10^6

# CO2
data_merge$CO2_dry_mmol=
  # (atm pressure * L CO2) / (R in L*atm/?K*mol * ?K temp) * 10^3 mmol/mol
  ((1*data_merge$CO2_dry_L)/(0.08206*(data_merge$avg_respT+273)))*10^3
```
 
Calculate chamber fluxes
```{r}
## Identify start of fluxes
Time_start <- meta_TMI[,c("fDOY_start","sample")]
flux.times=unique(Time_start$fDOY_start)

# Create new dataframes to hold final fluxes 
fluxes.CH4_may=data.frame(matrix(NA,nrow=length(flux.times)))
fluxes.CO2_may=data.frame(matrix(NA,nrow=length(flux.times)))

## Add named columns
# CH4
fluxes.CH4_may$Time_start=flux.times
fluxes.CH4_may$flux.CH4=0
fluxes.CH4_may$R2.CH4=0
fluxes.CH4_may$p.CH4=0

# CO2
fluxes.CO2_may$Time_start=flux.times
fluxes.CO2_may$flux.CO2=0
fluxes.CO2_may$R2.CO2=0
fluxes.CO2_may$p.CO2=0

## Remove initial empty column
fluxes.CO2_may=fluxes.CO2_may[,-1]
fluxes.CH4_may=fluxes.CH4_may[,-1]

## For each start time (may 2022)
for (i in flux.times) {
  ## CH4 ##
  # Subset data for one chamber measurement
  temp1=subset(data_merge,fDOY_start==i)
  # Set corresponding row of output table
  j=which(flux.times==i)
  # Determine if start time has a CH4 flux
  if (nrow(temp1)>0) {
    # If so:  
    # Calulate flux in umol/day using linear regression
    mod=with(temp1,lm(CH4_dry_umol~fDOY))
    # Save flux rate and R2 and p-value of slope in corresponding row of dataframe
    # flux rate
    fluxes.CH4_may$flux.CH4[j]=coef(mod)[2]/temp1$chamber_SA_m2 # chamber surface area
    # R2 of slope
    fluxes.CH4_may$R2.CH4[j]=summary(mod)$r.squared
    # p-value of slope
    fluxes.CH4_may$p.CH4[j]=summary(mod)$coefficients[2,4]
    # If not:
    # Fill rows of table with NA    
  } else {
    fluxes.CH4_may$flux.CH4[j]=NA
    fluxes.CH4_may$R2.CH4[j]=NA
    fluxes.CH4_may$p.CH4[j]=NA
  }
  ## CO2 ##
  # Subset data for one chamber measurement
  temp2=subset(data_merge,fDOY_start==i)
  # Calulate flux in mmol/day using linear regression
  mod=with(temp2,lm(CO2_dry_mmol~fDOY))
  # Save flux rate and R2 and p-value of slope in corresponding row of dataframe
  # flux rate
  fluxes.CO2_may$flux.CO2[j]=coef(mod)[2]/temp2$chamber_SA_m2 # chamber surface area
  # R2 of slope
  fluxes.CO2_may$R2.CO2[j]=summary(mod)$r.squared
  # p-value of slope
  fluxes.CO2_may$p.CO2[j]=summary(mod)$coefficients[2,4]
}
```

Finalize fluxes, merge back with metadata and extract
```{r}
# merge metadata back in to flux files 
# CO2
meta_TMI <- meta_TMI %>%
  dplyr::rename(Time_start = fDOY_start)

CO2_fluxfinal_may22 <- merge(fluxes.CO2_may, meta_TMI, by="Time_start")
CO2_fluxfinal_may22$campaign <- "may22"

# CH4
CH4_fluxfinal_may22 <- merge(fluxes.CH4_may, meta_TMI, by="Time_start")
CH4_fluxfinal_may22$campaign <- "may22"

# merge in species data to flux data 
species <- read.csv("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/supp_data/termite_ID.csv")
species <- species[c("X", "ID_cleaned")]

# merge sp and CH4flux_final, CO2_fluxfinal
colnames(species)[1] = "sample"
colnames(species)[2] = "species"

CH4_fluxfinal_may22 <- merge(CH4_fluxfinal_may22, species, by = c("sample")) 
CO2_fluxfinal_may22 <- merge(CO2_fluxfinal_may22, species, by = c("sample")) 

## Export fluxes as .csv file
# CO2 (units are mmol/day)
# write.csv(CO2_fluxfinal_may22,"/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/data_out/CO2_fluxfinal_may22.csv", row.names = FALSE)

# CH4 (units are umol/day)
# write.csv(CH4_fluxfinal_may22,"/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/data_out/CH4_fluxfinal_may22.csv", row.names = FALSE)
```

Preliminary visualization 
```{r}
# differences among species
ggplot(CH4_fluxfinal_may22, aes(x = species, y = flux.CH4, color = species))+
  geom_boxplot()+
  theme_classic()

# add time of day for measurements using GPS record
maytime <- read.csv("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/supp_data/may_mound_measure_times.csv")
maytime <- maytime[c("name", "realtime")]
names(maytime)[1] <- 'sample'
names(maytime)[2] <- 'flux_start_real'

CH4_fluxfinal_may22 <- merge(CH4_fluxfinal_may22, maytime, by = c("sample")) 
CH4_fluxfinal_may22 <- CH4_fluxfinal_may22 %>% relocate(flux_start_real, .after = Time_start)
CH4_fluxfinal_may22$flux_start_real <- as.POSIXct(CH4_fluxfinal_may22$flux_start_real, format = "%H:%M")

CO2_fluxfinal_may22 <- merge(CO2_fluxfinal_may22, maytime, by = c("sample")) 
CO2_fluxfinal_may22 <- CO2_fluxfinal_may22 %>% relocate(flux_start_real, .after = Time_start)
CO2_fluxfinal_may22$flux_start_real <- as.POSIXct(CO2_fluxfinal_may22$flux_start_real, format = "%H:%M")

# temperature throughout the day
ggplot(CH4_fluxfinal_may22, aes(x = flux_start_real, y = avg_respT))+
  geom_point()+
  theme_classic()

# flux and temperature relationship 
ggplot(CH4_fluxfinal_may22, aes(x = avg_respT, y = flux.CH4))+
  geom_point()+
  theme_classic()
```

QC check for all measurements
```{r}
# check r2 values 
# co2 
mean(CO2_fluxfinal_may22$R2.CO2)
ggplot(CO2_fluxfinal_may22, aes(R2.CO2))+
  geom_histogram()+
  theme_classic()

# ch4 
mean(CH4_fluxfinal_may22$R2.CH4)
ggplot(CH4_fluxfinal_may22, aes(R2.CH4))+
  geom_histogram()+
  theme_classic()

# check p values 
table(CO2_fluxfinal_may22$p.CO2<0.05)
table(CH4_fluxfinal_may22$p.CH4<0.05)

# extract everything with R2 < 0.4
CO2_fluxfinal_may22_fix <- CO2_fluxfinal_may22[CO2_fluxfinal_may22$R2.CO2 < 0.4,]
CH4_fluxfinal_may22_fix <- CH4_fluxfinal_may22[CH4_fluxfinal_may22$R2.CH4 < 0.4,]
```

QC CO2 
```{r}
# list of ID names by unique sample measurement (by creating sample-tag)
CO2_fluxfinal_may22_fix$ID_measurement <- paste(CO2_fluxfinal_may22_fix$sample, CO2_fluxfinal_may22_fix$tag, sep = "-")
data_merge$ID_measurement <- paste(data_merge$sample, data_merge$tag, sep = "-")
target_CO2 <- as.list(CO2_fluxfinal_may22_fix$ID_measurement)

# function to troubleshoot and visualize low correlation fluxes 
CO2_fix <- function(target_CO2, CO2_fluxfinal_may22_fix, data_merge){
  # select first ID name in list 
  target_ID <- target_CO2[1]
 
  # in fix file, set Time_start and fDOY_end as xlim_lower and xlim_upper 
  working <- CO2_fluxfinal_may22_fix[CO2_fluxfinal_may22_fix$ID_measurement == target_ID,]
  records <- data_merge[data_merge$ID_measurement == target_ID,]
  xlim_lower <- working$Time_start
  xlim_upper <- working$fDOY_end
  ylim_lower <- min(records$CO2_dry) - 5
  ylim_upper <- max(records$CO2_dry) + 5
  
  # plot and label with sample ID name - add to PDF where all are compiled
  ggplot(data_merge, aes(x = fDOY, y = CO2_dry))+
    geom_point()+
    theme_classic()+
    xlim(xlim_lower, xlim_upper)+
    ylim(ylim_lower, ylim_upper)+
    ggtitle(target_ID)
}

# test
out_CO2 <- sapply(seq_along(target_CO2), function(x) CO2_fix(target_CO2[[x]], CO2_fluxfinal_may22_fix, data_merge), USE.NAMES= TRUE, simplify = FALSE)

# ggsave(filename = "/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/troubleshooting_fluxes/plotCO2_may.pdf",
#    plot = marrangeGrob(out_CO2, nrow=1, ncol=1),
#    width = 15, height = 9)
```

QC CH4
```{r}
# list of ID names by unique sample measurement (by creating sample-tag)
CH4_fluxfinal_may22_fix$ID_measurement <- paste(CH4_fluxfinal_may22_fix$sample, CH4_fluxfinal_may22_fix$tag, sep = "-")
target_CH4 <- as.list(CH4_fluxfinal_may22_fix$ID_measurement)

# function to troubleshoot and visualize low correlation fluxes 
CH4_fix <- function(target_CH4, CH4_fluxfinal_may22_fix, data_merge){
  # select first ID name in list 
  target_ID <- target_CH4[1]
 
  # in fix file, set Time_start and fDOY_end as xlim_lower and xlim_upper 
  working <- CH4_fluxfinal_may22_fix[CH4_fluxfinal_may22_fix$ID_measurement == target_ID,]
  records <- data_merge[data_merge$ID_measurement == target_ID,]
  xlim_lower <- working$Time_start
  xlim_upper <- working$fDOY_end
  ylim_lower <- min(records$CH4_dry) - .1
  ylim_upper <- max(records$CH4_dry) + .1
  
  # plot and label with sample ID name - add to PDF where all are compiled
  ggplot(data_merge, aes(x = fDOY, y = CH4_dry))+
    geom_point()+
    theme_classic()+
    xlim(xlim_lower, xlim_upper)+
    ylim(ylim_lower, ylim_upper)+
    ggtitle(target_ID)
}

# test
out_CH4 <- sapply(seq_along(target_CH4), function(x) CH4_fix(target_CH4[[x]], CH4_fluxfinal_may22_fix, data_merge), USE.NAMES= TRUE, simplify = FALSE)

# ggsave(filename = "/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/troubleshooting_fluxes/plotCH4_may_rep.pdf",
#    plot = marrangeGrob(out_CH4, nrow=1, ncol=1),
#    width = 15, height = 9)
```

Fix QC flagged fluxes - same ones need to be fixed for CO2 and CH4
```{r}
# ones to remove all together 
remove <- c("MD19-1", "MD38-4", "MD39-3")

# display graphs of problem fluxes to define new start/endpoints
problem <- c("MD34-2", "MD45-3", "MD58-6")

ggplot(data_merge[data_merge$ID_measurement == problem[1],], aes(x = fDOY, y = CH4_dry))+
  geom_point()+
  theme_classic()+
  xlim(1065.4252, 1065.4262)

ggplot(data_merge[data_merge$ID_measurement == problem[2],], aes(x = fDOY, y = CH4_dry))+
  geom_point()+
  theme_classic()+
  xlim(1369.425, 1369.4258)

ggplot(data_merge[data_merge$ID_measurement == problem[3],], aes(x = fDOY, y = CH4_dry))+
  geom_point()+
  theme_classic()+
  xlim(1734.4312, 1734.432)

# read in troubleshooting metadata, select for problem fluxes 
trouble <- read.csv("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/troubleshooting_fluxes/troubleshooting_samples.csv")
trouble_test <- trouble[trouble$ID_measurement %in% problem,]
trouble_test <- trouble_test[1:3,]

# fix MD34-2
fix1 <- data_merge[data_merge$ID_measurement == problem[1],]
ggplot(fix1, aes(x = fDOY, y = CH4_dry))+
  geom_point()+
  theme_classic()

e <- trouble_test[1,]
new_end <- e$new_end
fixed1 <- fix1[!fix1$fDOY > new_end,]

ggplot(fixed1, aes(x = fDOY, y = CH4_dry))+
  geom_point()+
  theme_classic()

# fix MD45-3
fix2 <- data_merge[data_merge$ID_measurement == problem[2],]
ggplot(fix2, aes(x = fDOY, y = CH4_dry))+
  geom_point()+
  theme_classic()

e <- trouble_test[2,]
new_end <- e$new_end
fixed2 <- fix2[!fix2$fDOY > new_end,]

ggplot(fixed2, aes(x = fDOY, y = CH4_dry))+
  geom_point()+
  theme_classic()

# fix MD58-6
fix3 <- data_merge[data_merge$ID_measurement == problem[3],]
ggplot(fix3, aes(x = fDOY, y = CH4_dry))+
  geom_point()+
  theme_classic()

s <- trouble_test[3,]
new_start <- s$new_start
fixed3 <- fix3[!fix3$fDOY < new_start,]

ggplot(fixed3, aes(x = fDOY, y = CH4_dry))+
  geom_point()+
  theme_classic()

# now need to recalculate fluxes for fixed1 and fixed2 and fixed3
```

fixed fluxes
```{r}
# create fixed flux df 
fluxes.CH4_may_fix=data.frame(matrix(NA,nrow=3))
fluxes.CO2_may_fix=data.frame(matrix(NA,nrow=3))

## Add named columns
# CH4
fluxes.CH4_may_fix$Time_start=0
fluxes.CH4_may_fix$flux.CH4=0
fluxes.CH4_may_fix$R2.CH4=0
fluxes.CH4_may_fix$p.CH4=0

# CO2
fluxes.CO2_may_fix$Time_start=0
fluxes.CO2_may_fix$flux.CO2=0
fluxes.CO2_may_fix$R2.CO2=0
fluxes.CO2_may_fix$p.CO2=0

## Remove initial empty column
fluxes.CO2_may_fix=fluxes.CO2_may_fix[,-1]
fluxes.CH4_may_fix=fluxes.CH4_may_fix[,-1]

# CH4
# fixed1: MD34-2
# linear model and statistics
mod <- lm(CH4_dry_umol~fDOY, data = fixed1)
# Save flux rate and R2 and p-value of slope in corresponding row of dataframe
fluxes.CH4_may_fix[1,]$Time_start <- as.numeric(min(fixed1$fDOY))
fluxes.CH4_may_fix[1,]$flux.CH4=unique(coef(mod)[2]/fixed1$chamber_SA_m2) # chamber surface area
fluxes.CH4_may_fix[1,]$R2.CH4=summary(mod)$r.squared
fluxes.CH4_may_fix[1,]$p.CH4=summary(mod)$coefficients[2,4]

# fixed2: MD45-3
mod <- lm(CH4_dry_umol~fDOY, data = fixed2)
# Save flux rate and R2 and p-value of slope in corresponding row of dataframe
fluxes.CH4_may_fix[2,]$Time_start <- as.numeric(min(fixed2$fDOY))
fluxes.CH4_may_fix[2,]$flux.CH4=unique(coef(mod)[2]/fixed2$chamber_SA_m2) # chamber surface area
fluxes.CH4_may_fix[2,]$R2.CH4=summary(mod)$r.squared
fluxes.CH4_may_fix[2,]$p.CH4=summary(mod)$coefficients[2,4]

# fixed3: MD58-6
mod <- lm(CH4_dry_umol~fDOY, data = fixed3)
# Save flux rate and R2 and p-value of slope in corresponding row of dataframe
fluxes.CH4_may_fix[3,]$Time_start <- as.numeric(min(fixed3$fDOY))
fluxes.CH4_may_fix[3,]$flux.CH4=unique(coef(mod)[2]/fixed3$chamber_SA_m2) # chamber surface area
fluxes.CH4_may_fix[3,]$R2.CH4=summary(mod)$r.squared
fluxes.CH4_may_fix[3,]$p.CH4=summary(mod)$coefficients[2,4]

# CO2
# fixed1: MD34-2
# linear model and statistics
mod <- lm(CO2_dry_mmol~fDOY, data = fixed1)
# Save flux rate and R2 and p-value of slope in corresponding row of dataframe
fluxes.CO2_may_fix[1,]$Time_start <- as.numeric(min(fixed1$fDOY))
fluxes.CO2_may_fix[1,]$flux.CO2=unique(coef(mod)[2]/fixed1$chamber_SA_m2) # chamber surface area
fluxes.CO2_may_fix[1,]$R2.CO2=summary(mod)$r.squared
fluxes.CO2_may_fix[1,]$p.CO2=summary(mod)$coefficients[2,4]

# fixed2: MD45-3
mod <- lm(CO2_dry_mmol~fDOY, data = fixed2)
# Save flux rate and R2 and p-value of slope in corresponding row of dataframe
fluxes.CO2_may_fix[2,]$Time_start <- as.numeric(min(fixed2$fDOY))
fluxes.CO2_may_fix[2,]$flux.CO2=unique(coef(mod)[2]/fixed2$chamber_SA_m2) # chamber surface area
fluxes.CO2_may_fix[2,]$R2.CO2=summary(mod)$r.squared
fluxes.CO2_may_fix[2,]$p.CO2=summary(mod)$coefficients[2,4]

# fixed3: MD58-6
mod <- lm(CO2_dry_mmol~fDOY, data = fixed3)
# Save flux rate and R2 and p-value of slope in corresponding row of dataframe
fluxes.CO2_may_fix[3,]$Time_start <- as.numeric(min(fixed3$fDOY))
fluxes.CO2_may_fix[3,]$flux.CO2=unique(coef(mod)[2]/fixed3$chamber_SA_m2) # chamber surface area
fluxes.CO2_may_fix[3,]$R2.CO2=summary(mod)$r.squared
fluxes.CO2_may_fix[3,]$p.CO2=summary(mod)$coefficients[2,4]
```

add fixed fluxes to final df 
```{r}
# first pull out incorrect fluxes from original df 
a <- data_merge[data_merge$ID_measurement %in% problem,]
original_start_fix <- aggregate(fDOY ~ sample, a, function(x) min(x))
original_start_fix <-  original_start_fix[['fDOY']]

# add in corrected fluxes - ch4
working <- CH4_fluxfinal_may22[CH4_fluxfinal_may22$Time_start %in% original_start_fix,]
# remove so that corrected version can be merged back in 
CH4_fluxfinal_may22 <- CH4_fluxfinal_may22[!CH4_fluxfinal_may22$Time_start %in% original_start_fix,]

working[1,]$flux.CH4 <- fluxes.CH4_may_fix[1,]$flux.CH4
working[1,]$R2.CH4 <- fluxes.CH4_may_fix[1,]$R2.CH4
working[1,]$p.CH4 <- fluxes.CH4_may_fix[1,]$p.CH4

working[2,]$flux.CH4 <- fluxes.CH4_may_fix[2,]$flux.CH4
working[2,]$R2.CH4 <- fluxes.CH4_may_fix[2,]$R2.CH4
working[2,]$p.CH4 <- fluxes.CH4_may_fix[2,]$p.CH4

working[3,]$flux.CH4 <- fluxes.CH4_may_fix[3,]$flux.CH4
working[3,]$R2.CH4 <- fluxes.CH4_may_fix[3,]$R2.CH4
working[3,]$p.CH4 <- fluxes.CH4_may_fix[3,]$p.CH4

# add in corrected fluxes - co2
working_c <- CO2_fluxfinal_may22[CO2_fluxfinal_may22$Time_start %in% original_start_fix,]
# remove so that corrected version can be merged back in 
CO2_fluxfinal_may22 <- CO2_fluxfinal_may22[!CO2_fluxfinal_may22$Time_start %in% original_start_fix,]

working_c[1,]$flux.CO2 <- fluxes.CO2_may_fix[1,]$flux.CO2
working_c[1,]$R2.CO2 <- fluxes.CO2_may_fix[1,]$R2.CO2
working_c[1,]$p.CO2 <- fluxes.CO2_may_fix[1,]$p.CO2

working_c[2,]$flux.CO2 <- fluxes.CO2_may_fix[2,]$flux.CO2
working_c[2,]$R2.CO2 <- fluxes.CO2_may_fix[2,]$R2.CO2
working_c[2,]$p.CO2 <- fluxes.CO2_may_fix[2,]$p.CO2

working_c[3,]$flux.CO2 <- fluxes.CO2_may_fix[3,]$flux.CO2
working_c[3,]$R2.CO2 <- fluxes.CO2_may_fix[3,]$R2.CO2
working_c[3,]$p.CO2 <- fluxes.CO2_may_fix[3,]$p.CO2

# bring back together with finalized, fixed fluxes 
CH4_fluxfinal_may22_fixed <- rbind(CH4_fluxfinal_may22, working)
CO2_fluxfinal_may22_fixed <- rbind(CO2_fluxfinal_may22, working_c)

# remove the ones that were unfixable crap (3 traces)
CO2_fluxfinal_may22_fixed$ID_measurement <- paste(CO2_fluxfinal_may22_fixed$sample, CO2_fluxfinal_may22_fixed$tag, sep = "-")
CO2_fluxfinal_may22_fixed <- CO2_fluxfinal_may22_fixed[!CO2_fluxfinal_may22_fixed$ID_measurement %in% remove,]

CH4_fluxfinal_may22_fixed$ID_measurement <- paste(CH4_fluxfinal_may22_fixed$sample, CH4_fluxfinal_may22_fixed$tag, sep = "-")
CH4_fluxfinal_may22_fixed <- CH4_fluxfinal_may22_fixed[!CH4_fluxfinal_may22_fixed$ID_measurement %in% remove,]

# quick recheck on QC (anything that is <0.4 is due to near-zero fluxes)
# co2 
mean(CO2_fluxfinal_may22_fixed$R2.CO2)
ggplot(CO2_fluxfinal_may22_fixed, aes(R2.CO2))+
  geom_histogram()+
  theme_classic()

# ch4 
mean(CH4_fluxfinal_may22_fixed$R2.CH4)
ggplot(CH4_fluxfinal_may22_fixed, aes(R2.CH4))+
  geom_histogram()+
  theme_classic()

# fix true flux start times
CH4_fluxfinal_may22_fixed <- CH4_fluxfinal_may22_fixed[ , !names(CH4_fluxfinal_may22_fixed) %in% c("flux_start")]
colnames(CH4_fluxfinal_may22_fixed)[colnames(CH4_fluxfinal_may22_fixed) == "flux_start_real"] ="flux_start"

CO2_fluxfinal_may22_fixed <- CO2_fluxfinal_may22_fixed[ , !names(CO2_fluxfinal_may22_fixed) %in% c("flux_start")]
colnames(CO2_fluxfinal_may22_fixed)[colnames(CO2_fluxfinal_may22_fixed) == "flux_start_real"] ="flux_start"
```

set non-significant fluxes to 0 
```{r}
# pull out samples that have p < 0.05
table(CO2_fluxfinal_may22_fixed$p.CO2<0.05) # all CO2 is fine (either removed or fluxes are fixed)
table(CH4_fluxfinal_may22_fixed$p.CH4<0.05)
a <- CH4_fluxfinal_may22_fixed[CH4_fluxfinal_may22_fixed$p.CH4>0.05,]

# for CH4
CH4_to_zero <- a$ID_measurement
assign_zeros_CH4 <- c('flux.CH4', 'R2.CH4', 'p.CH4')

# function to reassign nonsignificant fluxes to 0 
zero_flux <- function(CH4_fluxfinal_may22_fixed, CH4_to_zero, assign_zeros_CH4) {
  # Replace columns in assign_zeros_CH4 with 0 for ID_measurement values in CH4_to_zero
  CH4_fluxfinal_may22_fixed <- CH4_fluxfinal_may22_fixed %>%
    mutate_at(vars(all_of(assign_zeros_CH4)), ~ ifelse(ID_measurement %in% CH4_to_zero, 0, .))
  
  return(CH4_fluxfinal_may22_fixed)
}

CH4_fluxfinal_may22_fixed_zeroadj <- zero_flux(CH4_fluxfinal_may22_fixed, CH4_to_zero, assign_zeros_CH4)
```

EXPORT
```{r}
## Export fluxes as .csv file
# CO2 (units are mmol/day)
# write.csv(CO2_fluxfinal_may22_fixed,"/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/finalfluxes/CO2_fluxfinal_may22_fixed.csv", row.names = FALSE)

# CH4 (units are umol/day)
# write.csv(CH4_fluxfinal_may22_fixed_zeroadj,"/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/finalfluxes/CH4_fluxfinal_may22_fixed.csv", row.names = FALSE)
```
