==> scripts/2.NOV22_process.Rmd <==
---
title: "2.NOV22_process"
author: "abbey yatsko"
date: "2023-09-12"
output: html_document
---
libraries
```{r workspace prep}
library(data.table)
library(ggplot2)
library(ggpubr)
library(sqldf)
library(dplyr)
library(tidyverse)
library(stringr)
library(plyr)
library(gridExtra)
library(plotly)
```

Nov 2022 data load and prep
```{r}
# workplace setup 
setwd("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/LGR_files/nov22LGR")
filenames <- list.files(pattern='f000',full.names=T)

# read in LGR datafiles to list and go from .txt to .csv 
data_s <- lapply(filenames,read.csv,skip=1)
data_s <- rbindlist(data_s)
data_s <- subset(data_s, X.CH4._ppm!="NA")
as.character(data_s$Time)

date_time <- strptime(data_s$Time,format='%d/%m/%Y %H:%M:%S')

# formatting dates and times ----
# add year,month,day,JD,hour,min,sec columns to dataframe
Year <- as.numeric(format(date_time,'%Y'))
Month <- as.numeric(format(date_time,'%m'))
Day <- as.numeric(format(date_time,'%d'))
fDOY <- as.numeric(julian(date_time,'2022-01-01'))  #Change for year
Hour <- as.numeric(format(date_time,'%k'))
Min <- as.numeric(format(date_time,'%M'))
Sec <- as.numeric(format(date_time,'%S'))
data_clean2 <- cbind(date_time,Year,Month,Day,fDOY,Hour,Min,Sec,data_s)

# save LGR data as dt
data_clean2 <- data.table(data_clean2)
```

Preliminary visualization for entire df (check and make sure fluxes show up)
```{r}
# for co2
co2 <- ggplot(data_clean2, aes(date_time, X.CO2.d_ppm)) +
  geom_point() +
  theme(axis.text.x = element_text(angle = 90, hjust = 1)) +
  ggtitle("CO2 flux")

# for ch4 
ch4 <- ggplot(data_clean2, aes(date_time, X.CH4.d_ppm)) +
  geom_point() +
  theme(axis.text.x = element_text(angle = 90, hjust = 1)) +
  ggtitle("CH4 flux")
```

Merge with metadata
```{r}
# read in file with metadata
meta_TMI2 <- read.csv('/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/LGR_metadata/TMI_flux_metadata_nov22.csv')

as.Date(meta_TMI2$measurement_day_actual, '%m/%d/%Y')
```

Add start and end time for flux measurements
```{r}
## Pull out date and time data
date_time_start <- strptime(paste(meta_TMI2$measurement_day_actual,meta_TMI2$flux_start),'%m/%d/%Y %H:%M:%S')

## Add 30 seconds to set buffered start time
date_time_start <- date_time_start+30

## Add 90 sec to determine end time
date_time_end <- date_time_start+90

## Add fDOY columns for start and endtime to dataframe
fDOY_start <- as.numeric(julian(date_time_start,'2022-01-01'))  #Change for year
fDOY_end <- as.numeric(julian(date_time_end,'2022-01-01'))  #Change for year
meta_TMI2 <- cbind(fDOY_start,fDOY_end,meta_TMI2)

# calculate average respiration temperature - add column
meta_TMI2$avg_respT <- rowMeans(meta_TMI2[ , c(9:11)], na.rm = TRUE)

# change sample to factor level variables 
meta_TMI2$sample <- as.factor(meta_TMI2$sample) 

# rename co2 and ch4 columns because naming format including '.' messes up the sql code 
data_clean2 <- data_clean2 %>%
  dplyr::rename(CH4 = X.CH4._ppm,
         CO2 = X.CO2._ppm,
         CH4_dry = X.CH4.d_ppm,
         CO2_dry = X.CO2.d_ppm)
```

Merge LGR files with metadata 
```{r}
# complete merge for respiration data and metadata (nov)
data_merge2 <- sqldf("select data_clean2.fDOY, data_clean2.CO2_dry, data_clean2.CH4_dry, meta_TMI2.sample, meta_TMI2.avg_respT, meta_TMI2.flux_source, meta_TMI2.position, meta_TMI2.tag, meta_TMI2.directon, meta_TMI2.chamber, meta_TMI2.chamber_SA_cm2, meta_TMI2.fDOY_start 
from data_clean2 LEFT JOIN meta_TMI2 ON (data_clean2.fDOY BETWEEN meta_TMI2.fDOY_start AND meta_TMI2.fDOY_end)")

# use 'complete.cases'
data_merge2 <- data_merge2[complete.cases(data_merge2), ]

# check to make sure all of the samples are there (24 total)
summary(unique(data_merge2$sample))
```

Visually check clipped bits of data 
```{r}
# co2 (xlim=c(317.5019, 317.5148) corresponds to only MD15 first file)
par(mfrow=c(2,1),mar=c(4,4,1,1))
with(data_clean2,plot(fDOY,CO2_dry,ylim=c(400,1300), xlim=c(317.5019, 317.5148)))
with(data_merge2,plot(fDOY,CO2_dry,ylim=c(400,1300), xlim=c(317.5019, 317.5148)))

# ch4 
par(mfrow=c(2,1),mar=c(4,4,1,1))
with(data_clean2,plot(fDOY,CH4_dry,ylim=c(1, 10), xlim=c(317.5019, 317.5148)))
with(data_merge2,plot(fDOY,CH4_dry,ylim=c(1,10), xlim=c(317.5019, 317.5148)))
```

Convert ppm to moles for CH4 and CO2
```{r}
# use known volume of chamber (in m3) to convert ppm of gas to liters of gas
# small, h = 5.7 cm, d = 7.8 cm
sm <- pi*((7.8/2)^2)*5.7

# medium, h = 10.8 cm, d = 10.3 cm
med <- pi*((10.3/2)^2)*10.8

# large, h = 16.1 cm, d = 15.3 cm
lg <- pi*((15.3/2)^2)*16.1

# mini, v = 119 ml
mini <- 119

volume_cm3 <- c(sm, med, lg, mini)
chamber <- c("sm", "med", "lg", "mini")
chamber_vols <- data.frame(volume_cm3, chamber)

# chamber volumes are in cm3 - convert to m3 for the next step 
chamber_vols$volume_m3 <- chamber_vols$volume_cm3*0.000001

data_merge2 <- merge(data_merge2, chamber_vols, by = c("chamber")) 
data_merge2 <- data_merge2 %>% relocate(chamber, .after = chamber_SA_cm2)

# also convert SA chamber to be in m2 (currently in cm2)
data_merge2$chamber_SA_m2 <- data_merge2$chamber_SA_cm2 * 0.0001
data_merge2 <- data_merge2 %>% relocate(chamber_SA_m2, .after = chamber_SA_cm2)

# CH4
data_merge2$CH4_dry_L=
  # parts CH4 per million parts air * volume of air in chamber (m3) * 1000 L per m3
  (data_merge2$CH4_dry/1000000) * data_merge2$volume_m3 * 1000 

# CO2
data_merge2$CO2_dry_L=
  # parts CO2 per million parts air * volume of air in chamber (m3) * 1000 L per m3
  (data_merge2$CO2_dry/1000000) * data_merge2$volume_m3 * 1000

# Use ideal gas law to calculate umol of CH4 or mmol of CO2
# CH4
data_merge2$CH4_dry_umol=
  # (atm pressure * L CH4) / (R in L*atm/?K*mol * ?K temp) * 10^6 umol/mol
  ((1*data_merge2$CH4_dry_L)/(0.08206*(data_merge2$avg_respT+273)))*10^6

# CO2
data_merge2$CO2_dry_mmol=
  # (atm pressure * L CO2) / (R in L*atm/?K*mol * ?K temp) * 10^3 mmol/mol
  ((1*data_merge2$CO2_dry_L)/(0.08206*(data_merge2$avg_respT+273)))*10^3
```

Calculate chamber fluxes
```{r}
## Identify start of fluxes
Time_start2 <- meta_TMI2[,c("fDOY_start","sample")]
flux.times2=unique(Time_start2$fDOY_start)

# Create new dataframes to hold final fluxes 
fluxes.CH4_nov=data.frame(matrix(NA,nrow=length(flux.times2)))
fluxes.CO2_nov=data.frame(matrix(NA,nrow=length(flux.times2)))

## Add named columns
# CH4
fluxes.CH4_nov$Time_start=flux.times2
fluxes.CH4_nov$flux.CH4=0
fluxes.CH4_nov$R2.CH4=0
fluxes.CH4_nov$p.CH4=0

# CO2
fluxes.CO2_nov$Time_start=flux.times2
fluxes.CO2_nov$flux.CO2=0
fluxes.CO2_nov$R2.CO2=0
fluxes.CO2_nov$p.CO2=0

## Remove initial empty column
fluxes.CO2_nov=fluxes.CO2_nov[,-1]
fluxes.CH4_nov=fluxes.CH4_nov[,-1]

## For each start time (nov 2022)
for (i in flux.times2) {
  ## CH4 ##
  # Subset data for one chamber measurement
  temp1=subset(data_merge2,fDOY_start==i)
  # Set corresponding row of output table
  j=which(flux.times2==i)
  # Determine if start time has a CH4 flux
  if (nrow(temp1)>0) {
    # If so:  
    # Calulate flux in umol/day using linear regression
    mod=with(temp1,lm(CH4_dry_umol~fDOY))
    # Save flux rate and R2 and p-value of slope in corresponding row of dataframe
    # flux rate
    fluxes.CH4_nov$flux.CH4[j]=coef(mod)[2]/temp1$chamber_SA_m2 # chamber surface area
    # R2 of slope
    fluxes.CH4_nov$R2.CH4[j]=summary(mod)$r.squared
    # p-value of slope
    fluxes.CH4_nov$p.CH4[j]=summary(mod)$coefficients[2,4]
    # If not:
    # Fill rows of table with NA    
  } else {
    fluxes.CH4_nov$flux.CH4[j]=NA
    fluxes.CH4_nov$R2.CH4[j]=NA
    fluxes.CH4_nov$p.CH4[j]=NA
  }
  ## CO2 ##
  # Subset data for one chamber measurement
  temp2=subset(data_merge2,fDOY_start==i)
  # Calulate flux in mmol/day using linear regression
  mod=with(temp2,lm(CO2_dry_mmol~fDOY))
  # Save flux rate and R2 and p-value of slope in corresponding row of dataframe
  # flux rate
  fluxes.CO2_nov$flux.CO2[j]=coef(mod)[2]/temp2$chamber_SA_m2 # chamber surface area
  # R2 of slope
  fluxes.CO2_nov$R2.CO2[j]=summary(mod)$r.squared
  # p-value of slope
  fluxes.CO2_nov$p.CO2[j]=summary(mod)$coefficients[2,4]
}
```

Finalize fluxes, merge back with metadata and extract
```{r}
# merge metadata back in to flux files 
# CO2
meta_TMI2 <- meta_TMI2 %>%
  dplyr::rename(Time_start = fDOY_start)

CO2_fluxfinal_nov22 <- merge(fluxes.CO2_nov, meta_TMI2, by="Time_start")
CO2_fluxfinal_nov22$campaign <- "nov22"

# CH4
CH4_fluxfinal_nov22 <- merge(fluxes.CH4_nov, meta_TMI2, by="Time_start")
CH4_fluxfinal_nov22$campaign <- "nov22"

# merge in species data to flux data 
species <- read.csv("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/supp_data/termite_ID.csv")
species <- species[c("X", "ID_cleaned")]

# merge sp and CH4flux_final, CO2_fluxfinal
colnames(species)[1] = "sample"
colnames(species)[2] = "species"

CH4_fluxfinal_nov22 <- merge(CH4_fluxfinal_nov22, species, by = c("sample")) 
CO2_fluxfinal_nov22 <- merge(CO2_fluxfinal_nov22, species, by = c("sample")) 

## Export fluxes as .csv file
# CO2 (units are mmol/day)
# write.csv(CO2_fluxfinal_nov22,"/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/data_out/CO2_fluxfinal_nov22.csv", row.names = FALSE)

# CH4 (units are umol/day)
# write.csv(CH4_fluxfinal_nov22,"/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/data_out/CH4_fluxfinal_nov22.csv", row.names = FALSE)
```

Preliminary visualization 
```{r}
# differences among species
ggplot(CH4_fluxfinal_nov22, aes(x = species, y = flux.CH4, color = species))+
  geom_boxplot()+
  theme_classic()

# temperature throughout the day 
CH4_fluxfinal_nov22$flux_start <- as.POSIXct(CH4_fluxfinal_nov22$flux_start, format = "%H:%M:%S")

ggplot(CH4_fluxfinal_nov22, aes(x = flux_start, y = avg_respT))+
  geom_point()+
  theme_classic()

# flux and temperature relationship 
ggplot(CH4_fluxfinal_nov22, aes(x = avg_respT, y = flux.CH4))+
  geom_point()+
  theme_classic()
```

QC check for all measurements
```{r}
# check r2 values 
# co2 
mean(CO2_fluxfinal_nov22$R2.CO2)
ggplot(CO2_fluxfinal_nov22, aes(R2.CO2))+
  geom_histogram()+
  theme_classic()

# ch4 
mean(CH4_fluxfinal_nov22$R2.CH4)
ggplot(CH4_fluxfinal_nov22, aes(R2.CH4))+
  geom_histogram()+
  theme_classic()

# check p values 
table(CO2_fluxfinal_nov22$p.CO2<0.05)
table(CH4_fluxfinal_nov22$p.CH4<0.05)

# extract everything with R2 < 0.4
CO2_fluxfinal_nov22_fix <- CO2_fluxfinal_nov22[CO2_fluxfinal_nov22$R2.CO2 < 0.4,]
CH4_fluxfinal_nov22_fix <- CH4_fluxfinal_nov22[CH4_fluxfinal_nov22$R2.CH4 < 0.4,]
```

QC CO2 
```{r}
# list of ID names by unique sample measurement (by creating sample-tag)
CO2_fluxfinal_nov22_fix$ID_measurement <- paste(CO2_fluxfinal_nov22_fix$sample, CO2_fluxfinal_nov22_fix$tag, sep = "-")
data_merge2$ID_measurement <- paste(data_merge2$sample, data_merge2$tag, sep = "-")
target_CO2 <- as.list(CO2_fluxfinal_nov22_fix$ID_measurement)

# function to troubleshoot and visualize low correlation fluxes 
CO2_fix <- function(target_CO2, CO2_fluxfinal_nov22_fix, data_merge2){
  # select first ID name in list 
  target_ID <- target_CO2[1]
 
  # in fix file, set Time_start and fDOY_end as xlim_lower and xlim_upper 
  working <- CO2_fluxfinal_nov22_fix[CO2_fluxfinal_nov22_fix$ID_measurement == target_ID,]
  records <- data_merge2[data_merge2$ID_measurement == target_ID,]
  xlim_lower <- working$Time_start
  xlim_upper <- working$fDOY_end
  ylim_lower <- min(records$CO2_dry) - 5
  ylim_upper <- max(records$CO2_dry) + 5
  
  # plot and label with sample ID name - add to PDF where all are compiled
  ggplot(data_merge2, aes(x = fDOY, y = CO2_dry))+
    geom_point()+
    theme_classic()+
    xlim(xlim_lower, xlim_upper)+
    ylim(ylim_lower, ylim_upper)+
    ggtitle(target_ID)
}

# test
out_CO2 <- sapply(seq_along(target_CO2), function(x) CO2_fix(target_CO2[[x]], CO2_fluxfinal_nov22_fix, data_merge2), USE.NAMES= TRUE, simplify = FALSE)

# ggsave(filename = "/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/troubleshooting_fluxes/plotCO2_nov.pdf",
#    plot = marrangeGrob(out_CO2, nrow=1, ncol=1),
#    width = 15, height = 9)
```

QC CH4
```{r}
# list of ID names by unique sample measurement (by creating sample-tag)
CH4_fluxfinal_nov22_fix$ID_measurement <- paste(CH4_fluxfinal_nov22_fix$sample, CH4_fluxfinal_nov22_fix$tag, sep = "-")
target_CH4 <- as.list(CH4_fluxfinal_nov22_fix$ID_measurement)

# function to troubleshoot and visualize low correlation fluxes 
CH4_fix <- function(target_CH4, CH4_fluxfinal_nov22_fix, data_merge2){
  # select first ID name in list 
  target_ID <- target_CH4[1]
 
  # in fix file, set Time_start and fDOY_end as xlim_lower and xlim_upper 
  working <- CH4_fluxfinal_nov22_fix[CH4_fluxfinal_nov22_fix$ID_measurement == target_ID,]
  records <- data_merge2[data_merge2$ID_measurement == target_ID,]
  xlim_lower <- working$Time_start
  xlim_upper <- working$fDOY_end
  ylim_lower <- min(records$CH4_dry) - .1
  ylim_upper <- max(records$CH4_dry) + .1
  
  # plot and label with sample ID name - add to PDF where all are compiled
  ggplot(data_merge2, aes(x = fDOY, y = CH4_dry))+
    geom_point()+
    theme_classic()+
    xlim(xlim_lower, xlim_upper)+
    ylim(ylim_lower, ylim_upper)+
    ggtitle(target_ID)
}

# test
out_CH4 <- sapply(seq_along(target_CH4), function(x) CH4_fix(target_CH4[[x]], CH4_fluxfinal_nov22_fix, data_merge2), USE.NAMES= TRUE, simplify = FALSE)

# ggsave(filename = "/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/troubleshooting_fluxes/plotCH4_nov.pdf",
#    plot = marrangeGrob(out_CH4, nrow=1, ncol=1),
#    width = 15, height = 9)
```

Fix QC flagged fluxes 
```{r}
# display graphs of problem fluxes to define new start/endpoints
problem <- c("MD10-3", "MD55-6")

p1 <- ggplot(data_merge2[data_merge2$ID_measurement == problem[1],], aes(x = fDOY, y = CO2_dry))+
  geom_point()+
  theme_classic()
ggplotly(p1)

p2 <- ggplot(data_merge2[data_merge2$ID_measurement == problem[2],], aes(x = fDOY, y = CO2_dry))+
  geom_point()+
  theme_classic()
ggplotly(p2)

# read in troubleshooting metadata, select for problem fluxes 
trouble <- read.csv("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/troubleshooting_fluxes/troubleshooting_samples.csv")
trouble_test <- trouble[trouble$ID_measurement %in% problem,]
trouble_test <- trouble_test[1:2,]

# fix MD10-3
fix1 <- data_merge2[data_merge2$ID_measurement == problem[1],]
ggplot(fix1, aes(x = fDOY, y = CO2_dry))+
  geom_point()+
  theme_classic()

s <- trouble_test[1,]
new_start <- s$new_start
fixed1 <- fix1[!fix1$fDOY < new_start,]

ggplot(fixed1, aes(x = fDOY, y = CO2_dry))+
  geom_point()+
  theme_classic()

# fix MD55-6
fix2 <- data_merge2[data_merge2$ID_measurement == problem[2],]
ggplot(fix2, aes(x = fDOY, y = CH4_dry))+
  geom_point()+
  theme_classic()

s <- trouble_test[2,]
new_start <- s$new_start
fixed2 <- fix2[!fix2$fDOY < new_start,]

ggplot(fixed2, aes(x = fDOY, y = CH4_dry))+
  geom_point()+
  theme_classic()

# now need to calculate fluxes for fixed1 and fixed2
```

fixed fluxes
```{r}
# create fixed flux df 
fluxes.CH4_nov_fix=data.frame(matrix(NA,nrow=2))
fluxes.CO2_nov_fix=data.frame(matrix(NA,nrow=2))

## Add named columns
# CH4
fluxes.CH4_nov_fix$Time_start=0
fluxes.CH4_nov_fix$flux.CH4=0
fluxes.CH4_nov_fix$R2.CH4=0
fluxes.CH4_nov_fix$p.CH4=0

# CO2
fluxes.CO2_nov_fix$Time_start=0
fluxes.CO2_nov_fix$flux.CO2=0
fluxes.CO2_nov_fix$R2.CO2=0
fluxes.CO2_nov_fix$p.CO2=0

## Remove initial empty column
fluxes.CO2_nov_fix=fluxes.CO2_nov_fix[,-1]
fluxes.CH4_nov_fix=fluxes.CH4_nov_fix[,-1]

# CH4
# fixed1: MD10-3
# linear model and statistics
mod <- lm(CH4_dry_umol~fDOY, data = fixed1)
# Save flux rate and R2 and p-value of slope in corresponding row of dataframe
fluxes.CH4_nov_fix[1,]$Time_start <- as.numeric(min(fixed1$fDOY))
fluxes.CH4_nov_fix[1,]$flux.CH4=unique(coef(mod)[2]/fixed1$chamber_SA_m2) # chamber surface area
fluxes.CH4_nov_fix[1,]$R2.CH4=summary(mod)$r.squared
fluxes.CH4_nov_fix[1,]$p.CH4=summary(mod)$coefficients[2,4]

# fixed2: MD55-6
mod <- lm(CH4_dry_umol~fDOY, data = fixed2)

# Save flux rate and R2 and p-value of slope in corresponding row of dataframe
fluxes.CH4_nov_fix[2,]$Time_start <- as.numeric(min(fixed2$fDOY))
fluxes.CH4_nov_fix[2,]$flux.CH4=unique(coef(mod)[2]/fixed2$chamber_SA_m2) # chamber surface area
fluxes.CH4_nov_fix[2,]$R2.CH4=summary(mod)$r.squared
fluxes.CH4_nov_fix[2,]$p.CH4=summary(mod)$coefficients[2,4]

# CO2
# fixed1: MD10-3
# linear model and statistics
mod <- lm(CO2_dry_mmol~fDOY, data = fixed1)
# Save flux rate and R2 and p-value of slope in corresponding row of dataframe
fluxes.CO2_nov_fix[1,]$Time_start <- as.numeric(min(fixed1$fDOY))
fluxes.CO2_nov_fix[1,]$flux.CO2=unique(coef(mod)[2]/fixed1$chamber_SA_m2) # chamber surface area
fluxes.CO2_nov_fix[1,]$R2.CO2=summary(mod)$r.squared
fluxes.CO2_nov_fix[1,]$p.CO2=summary(mod)$coefficients[2,4]

# fixed2: MD55-6
mod <- lm(CO2_dry_mmol~fDOY, data = fixed2)
# Save flux rate and R2 and p-value of slope in corresponding row of dataframe
fluxes.CO2_nov_fix[2,]$Time_start <- as.numeric(min(fixed2$fDOY))
fluxes.CO2_nov_fix[2,]$flux.CO2=unique(coef(mod)[2]/fixed2$chamber_SA_m2) # chamber surface area
fluxes.CO2_nov_fix[2,]$R2.CO2=summary(mod)$r.squared
fluxes.CO2_nov_fix[2,]$p.CO2=summary(mod)$coefficients[2,4]
```

add fixed fluxes to final df 
```{r}
# first pull out incorrect fluxes from original df 
a <- data_merge2[data_merge2$ID_measurement %in% problem,]
original_start_fix <- aggregate(fDOY ~ sample, a, function(x) min(x))
original_start_fix <-  original_start_fix[['fDOY']]

# add in corrected fluxes - ch4
working <- CH4_fluxfinal_nov22[CH4_fluxfinal_nov22$Time_start %in% original_start_fix,]
# remove so that corrected version can be merged back in 
CH4_fluxfinal_nov22 <- CH4_fluxfinal_nov22[!CH4_fluxfinal_nov22$Time_start %in% original_start_fix,]

working[1,]$flux.CH4 <- fluxes.CH4_nov_fix[1,]$flux.CH4
working[1,]$R2.CH4 <- fluxes.CH4_nov_fix[1,]$R2.CH4
working[1,]$p.CH4 <- fluxes.CH4_nov_fix[1,]$p.CH4

working[2,]$flux.CH4 <- fluxes.CH4_nov_fix[2,]$flux.CH4
working[2,]$R2.CH4 <- fluxes.CH4_nov_fix[2,]$R2.CH4
working[2,]$p.CH4 <- fluxes.CH4_nov_fix[2,]$p.CH4

# add in corrected fluxes - co2
working_c <- CO2_fluxfinal_nov22[CO2_fluxfinal_nov22$Time_start %in% original_start_fix,]
# remove so that corrected version can be merged back in 
CO2_fluxfinal_nov22 <- CO2_fluxfinal_nov22[!CO2_fluxfinal_nov22$Time_start %in% original_start_fix,]

working_c[1,]$flux.CO2 <- fluxes.CO2_nov_fix[1,]$flux.CO2
working_c[1,]$R2.CO2 <- fluxes.CO2_nov_fix[1,]$R2.CO2
working_c[1,]$p.CO2 <- fluxes.CO2_nov_fix[1,]$p.CO2

working_c[2,]$flux.CO2 <- fluxes.CO2_nov_fix[2,]$flux.CO2
working_c[2,]$R2.CO2 <- fluxes.CO2_nov_fix[2,]$R2.CO2
working_c[2,]$p.CO2 <- fluxes.CO2_nov_fix[2,]$p.CO2

# bring back together with finalized, fixed fluxes 
CH4_fluxfinal_nov22_fixed <- rbind(CH4_fluxfinal_nov22, working)
CO2_fluxfinal_nov22_fixed <- rbind(CO2_fluxfinal_nov22, working_c)

# define ID_measurement column 
CO2_fluxfinal_nov22_fixed$ID_measurement <- paste(CO2_fluxfinal_nov22_fixed$sample, CO2_fluxfinal_nov22_fixed$tag, sep = "-")
CH4_fluxfinal_nov22_fixed$ID_measurement <- paste(CH4_fluxfinal_nov22_fixed$sample, CH4_fluxfinal_nov22_fixed$tag, sep = "-")

# quick recheck on QC (anything that is <0.4 is due to near-zero fluxes)
# co2 
mean(CO2_fluxfinal_nov22_fixed$R2.CO2)
ggplot(CO2_fluxfinal_nov22_fixed, aes(R2.CO2))+
  geom_histogram()+
  theme_classic()

# ch4 
mean(CH4_fluxfinal_nov22_fixed$R2.CH4)
ggplot(CH4_fluxfinal_nov22_fixed, aes(R2.CH4))+
  geom_histogram()+
  theme_classic()
```

set non-significant fluxes to 0 
```{r}
# pull out samples that have p < 0.05
table(CO2_fluxfinal_nov22_fixed$p.CO2<0.05) # all CO2 is fine (either removed or fluxes are fixed)
table(CH4_fluxfinal_nov22_fixed$p.CH4<0.05)
a <- CH4_fluxfinal_nov22_fixed[CH4_fluxfinal_nov22_fixed$p.CH4>0.05,]

# for CH4
CH4_to_zero <- a$ID_measurement
assign_zeros_CH4 <- c('flux.CH4', 'R2.CH4', 'p.CH4')

# function to reassign nonsignificant fluxes to 0 
zero_flux <- function(CH4_fluxfinal_nov22_fixed, CH4_to_zero, assign_zeros_CH4) {
  # Replace columns in assign_zeros_CH4 with 0 for ID_measurement values in CH4_to_zero
  CH4_fluxfinal_nov22_fixed <- CH4_fluxfinal_nov22_fixed %>%
    mutate_at(vars(all_of(assign_zeros_CH4)), ~ ifelse(ID_measurement %in% CH4_to_zero, 0, .))
  
  return(CH4_fluxfinal_nov22_fixed)
}

CH4_fluxfinal_nov22_fixed_zeroadj <- zero_flux(CH4_fluxfinal_nov22_fixed, CH4_to_zero, assign_zeros_CH4)
```

EXPORT
```{r}
## Export fluxes as .csv file
# CO2 (units are mmol/day)
# write.csv(CO2_fluxfinal_nov22_fixed,"/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/finalfluxes/CO2_fluxfinal_nov22_fixed.csv", row.names = FALSE)

# CH4 (units are umol/day)
# write.csv(CH4_fluxfinal_nov22_fixed_zeroadj,"/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/finalfluxes/CH4_fluxfinal_nov22_fixed.csv", row.names = FALSE)
```
