==> scripts/5.DAILY_process.Rmd <==
---
title: "5.DAILY_process"
author: "abbey yatsko"
date: "2023-09-12"
output: html_document
---
libraries
```{r workspace prep}
library(data.table)
library(ggplot2)
library(ggpubr)
library(sqldf)
library(dplyr)
library(tidyverse)
library(stringr)
library(plyr)
library(plotly)
library(gridExtra)
library(lme4)
library(easystats)
library(respirometry)
```

Aug 2023 data load and prep
```{r}
setwd("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/LGR_files/aug23LGR_daily")
filenames <- list.files(pattern='f00',full.names=T)

# read in LGR datafiles to list and go from .txt to .csv 
data_s <- lapply(filenames,read.csv,skip=1)
data_s <- rbindlist(data_s)
data_s <- subset(data_s, X.CH4._ppm!="NA")
as.character(data_s$Time)

date_time <- strptime(data_s$Time,format='%d/%m/%Y %H:%M:%S')

# formatting dates and times - add year,month,day,JD,hour,min,sec columns to dataframe
Year <- as.numeric(format(date_time,'%Y'))
Month <- as.numeric(format(date_time,'%m'))
Day <- as.numeric(format(date_time,'%d'))
fDOY <- as.numeric(julian(date_time,'2023-01-01'))  #Change for year
Hour <- as.numeric(format(date_time,'%k'))
Min <- as.numeric(format(date_time,'%M'))
Sec <- as.numeric(format(date_time,'%S'))
data_clean_day <- cbind(date_time,Year,Month,Day,fDOY,Hour,Min,Sec,data_s)

# save LGR data as dt
data_clean_day <- data.table(data_clean_day)
```

Preliminary visualization for entire df (check and make sure fluxes show up)
```{r}
# for co2
co2 <- ggplot(data_clean_day, aes(date_time, X.CO2.d_ppm)) +
  geom_point() +
  theme(axis.text.x = element_text(angle = 90, hjust = 1)) +
  ggtitle("CO2 flux")

# for ch4 
ch4 <- ggplot(data_clean_day, aes(date_time, X.CH4.d_ppm)) +
  geom_point() +
  theme(axis.text.x = element_text(angle = 90, hjust = 1)) +
  ggtitle("CH4 flux")
```

Merge with metadata
```{r}
meta_TMI_day <- read.csv('/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/LGR_metadata/TMI_flux_metadata_aug_daily.csv')

as.Date(meta_TMI_day$measurement_day_actual, '%m/%d/%Y')
```

Add start and end time for flux measurements
```{r}
## Pull out date and time data
date_time_start <- strptime(paste(meta_TMI_day$measurement_day_actual,meta_TMI_day$flux_start),'%m/%d/%Y %H:%M:%S')

## Add 30 seconds to set buffered start time
date_time_start <- date_time_start+30

## Add 90 sec to determine end time
date_time_end <- date_time_start+90

## Add fDOY columns for start and endtime to dataframe
fDOY_start <- as.numeric(julian(date_time_start,'2023-01-01'))  #Change for year
fDOY_end <- as.numeric(julian(date_time_end,'2023-01-01'))  #Change for year
meta_TMI_day <- cbind(fDOY_start,fDOY_end,meta_TMI_day)

# calculate average respiration temperature - add column
meta_TMI_day$avg_respT <- rowMeans(meta_TMI_day[ , c(9:11)], na.rm = TRUE)

# change sample to factor level variables 
meta_TMI_day$sample <- as.factor(meta_TMI_day$sample) 

# rename co2 and ch4 columns because naming format including '.' messes up the sql code 
data_clean_day <- data_clean_day %>%
  dplyr::rename(CH4 = X.CH4._ppm,
                CO2 = X.CO2._ppm,
                CH4_dry = X.CH4.d_ppm,
                CO2_dry = X.CO2.d_ppm)
```

Merge LGR files with metadata 
```{r}
# complete merge for respiration data and metadata (aug)
data_merge_day <- sqldf("select data_clean_day.fDOY, data_clean_day.CO2_dry, data_clean_day.CH4_dry, meta_TMI_day.sample, meta_TMI_day.avg_respT, meta_TMI_day.flux_source, meta_TMI_day.position, meta_TMI_day.tag, meta_TMI_day.directon, meta_TMI_day.chamber, meta_TMI_day.chamber_SA_cm2, meta_TMI_day.time_of_day, meta_TMI_day.fDOY_start 
from data_clean_day LEFT JOIN meta_TMI_day ON (data_clean_day.fDOY BETWEEN meta_TMI_day.fDOY_start AND meta_TMI_day.fDOY_end)")

# use 'complete.cases'
data_merge_day <- data_merge_day[complete.cases(data_merge_day), ]

# check to make sure all of the samples are there (9 mounds total)
summary(unique(data_merge_day$sample))
```

Visually check clipped bits of data 
```{r}
# co2 (xlim=c(226.4767, 226.4833) corresponds to only DM1, first measurement)
par(mfrow=c(2,1),mar=c(4,4,1,1))
with(data_clean_day,plot(fDOY,CO2_dry,ylim=c(400,1300), xlim=c(226.4767, 226.4833)))
with(data_merge_day,plot(fDOY,CO2_dry,ylim=c(400,1300), xlim=c(226.4767, 226.4833)))

# ch4 
par(mfrow=c(2,1),mar=c(4,4,1,1))
with(data_clean_day,plot(fDOY,CH4_dry,ylim=c(1, 10), xlim=c(226.4767, 226.4833)))
with(data_merge_day,plot(fDOY,CH4_dry,ylim=c(1,10), xlim=c(226.4767, 226.4833)))
```

Convert ppm to moles for CH4 and CO2
```{r}
# use known volume of chamber (in m3) to convert ppm of gas to liters of gas
# small, h = 5.7 cm, d = 7.8 cm
sm <- pi*((7.8/2)^2)*5.7

# medium, h = 10.8 cm, d = 10.3 cm
med <- pi*((10.3/2)^2)*10.8

# large, h = 16.1 cm, d = 15.3 cm
lg <- pi*((15.3/2)^2)*16.1

# mini, v = 119 ml
mini <- 119

volume_cm3 <- c(sm, med, lg, mini)
chamber <- c("sm", "med", "lg", "mini")
chamber_vols <- data.frame(volume_cm3, chamber)

# chamber volumes are in cm3 - convert to m3 for the next step 
chamber_vols$volume_m3 <- chamber_vols$volume_cm3*0.000001

data_merge_day <- merge(data_merge_day, chamber_vols, by = c("chamber")) 
data_merge_day <- data_merge_day %>% relocate(chamber, .after = chamber_SA_cm2)

# also convert SA chamber to be in m2 (currently in cm2)
data_merge_day$chamber_SA_m2 <- data_merge_day$chamber_SA_cm2 * 0.0001
data_merge_day <- data_merge_day %>% relocate(chamber_SA_m2, .after = chamber_SA_cm2)

# CH4
data_merge_day$CH4_dry_L=
  # parts CH4 per million parts air * volume of air in chamber (m3) * 1000 L per m3
  (data_merge_day$CH4_dry/1000000) * data_merge_day$volume_m3 * 1000 

# CO2
data_merge_day$CO2_dry_L=
  # parts CO2 per million parts air * volume of air in chamber (m3) * 1000 L per m3
  (data_merge_day$CO2_dry/1000000) * data_merge_day$volume_m3 * 1000 

# Use ideal gas law to calculate umol of CH4 or mmol of CO2
# CH4
data_merge_day$CH4_dry_umol=
  # (atm pressure * L CH4) / (R in L*atm/?K*mol * ?K temp) * 10^6 umol/mol
  ((1*data_merge_day$CH4_dry_L)/(0.08206*(data_merge_day$avg_respT+273)))*10^6

# CO2
data_merge_day$CO2_dry_mmol=
  # (atm pressure * L CO2) / (R in L*atm/?K*mol * ?K temp) * 10^3 mmol/mol
  ((1*data_merge_day$CO2_dry_L)/(0.08206*(data_merge_day$avg_respT+273)))*10^3
```

Calculate chamber fluxes
```{r}
## Identify start of fluxes
Time_start <- meta_TMI_day[,c("fDOY_start","sample")]
flux.times=unique(Time_start$fDOY_start)

# Create new dataframes to hold final fluxes 
fluxes.CH4_aug=data.frame(matrix(NA,nrow=length(flux.times)))
fluxes.CO2_aug=data.frame(matrix(NA,nrow=length(flux.times)))

## Add named columns
# CH4
fluxes.CH4_aug$Time_start=flux.times
fluxes.CH4_aug$flux.CH4=0
fluxes.CH4_aug$R2.CH4=0
fluxes.CH4_aug$p.CH4=0

# CO2
fluxes.CO2_aug$Time_start=flux.times
fluxes.CO2_aug$flux.CO2=0
fluxes.CO2_aug$R2.CO2=0
fluxes.CO2_aug$p.CO2=0

## Remove initial empty column
fluxes.CO2_aug=fluxes.CO2_aug[,-1]
fluxes.CH4_aug=fluxes.CH4_aug[,-1]

## For each start time (aug 2023)
for (i in flux.times) {
  ## CH4 ##
  # Subset data for one chamber measurement
  temp1=subset(data_merge_day,fDOY_start==i)
  # Set corresponding row of output table
  j=which(flux.times==i)
  # Determine if start time has a CH4 flux
  if (nrow(temp1)>0) {
    # If so:  
    # Calulate flux in umol/day using linear regression
    mod=with(temp1,lm(CH4_dry_umol~fDOY))
    # Save flux rate and R2 and p-value of slope in corresponding row of dataframe
    # flux rate
    fluxes.CH4_aug$flux.CH4[j]=coef(mod)[2]/temp1$chamber_SA_m2 # chamber surface area
    # R2 of slope
    fluxes.CH4_aug$R2.CH4[j]=summary(mod)$r.squared
    # p-value of slope
    fluxes.CH4_aug$p.CH4[j]=summary(mod)$coefficients[2,4]
    # If not:
    # Fill rows of table with NA    
  } else {
    fluxes.CH4_aug$flux.CH4[j]=NA
    fluxes.CH4_aug$R2.CH4[j]=NA
    fluxes.CH4_aug$p.CH4[j]=NA
  }
  ## CO2 ##
  # Subset data for one chamber measurement
  temp2=subset(data_merge_day,fDOY_start==i)
  # Calulate flux in mmol/day using linear regression
  mod=with(temp2,lm(CO2_dry_mmol~fDOY))
  # Save flux rate and R2 and p-value of slope in corresponding row of dataframe
  # flux rate
  fluxes.CO2_aug$flux.CO2[j]=coef(mod)[2]/temp2$chamber_SA_m2 # chamber surface area
  # R2 of slope
  fluxes.CO2_aug$R2.CO2[j]=summary(mod)$r.squared
  # p-value of slope
  fluxes.CO2_aug$p.CO2[j]=summary(mod)$coefficients[2,4]
}
```

Finalize fluxes, merge back with metadata and extract
```{r}
# merge metadata back in to flux files 
# CO2
meta_TMI_day <- meta_TMI_day %>%
  dplyr::rename(Time_start = fDOY_start)

CO2_fluxfinal_aug23_day <- merge(fluxes.CO2_aug, meta_TMI_day, by="Time_start")
CO2_fluxfinal_aug23_day$campaign <- "aug23"

# CH4
CH4_fluxfinal_aug23_day <- merge(fluxes.CH4_aug, meta_TMI_day, by="Time_start")
CH4_fluxfinal_aug23_day$campaign <- "aug23"

# merge in species data to flux data 
species <- read.csv("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/supp_data/daily_species.csv")
species <- species[c("X", "ID_cleaned")]

# merge sp and CH4flux_final, CO2_fluxfinal
colnames(species)[1] = "sample"
colnames(species)[2] = "species"

CH4_fluxfinal_aug23_day <- merge(CH4_fluxfinal_aug23_day, species, by = c("sample")) 
CO2_fluxfinal_aug23_day <- merge(CO2_fluxfinal_aug23_day, species, by = c("sample")) 

## Export fluxes as .csv file
# CO2 (units are mmol/day)
# write.csv(CO2_fluxfinal_aug23_day,"/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/data_out/CO2_fluxfinal_aug23_day.csv", row.names = FALSE)

# CH4 (units are umol/day)
# write.csv(CH4_fluxfinal_aug23_day,"/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/data_out/CH4_fluxfinal_aug23_day.csv", row.names = FALSE)
```

QC check for all measurements
```{r}
# check r2 values 
# co2 
mean(CO2_fluxfinal_aug23_day$R2.CO2)
ggplot(CO2_fluxfinal_aug23_day, aes(R2.CO2))+
  geom_histogram()+
  theme_classic()

# ch4 
mean(CH4_fluxfinal_aug23_day$R2.CH4)
ggplot(CH4_fluxfinal_aug23_day, aes(R2.CH4))+
  geom_histogram()+
  theme_classic()

# check p values
table(CO2_fluxfinal_aug23_day$p.CO2<0.05)
table(CH4_fluxfinal_aug23_day$p.CH4<0.05)

# extract everything with R2 < 0.4
CO2_fluxfinal_aug23_fix <- CO2_fluxfinal_aug23_day[CO2_fluxfinal_aug23_day$R2.CO2 < 0.4,]
CH4_fluxfinal_aug23_fix <- CH4_fluxfinal_aug23_day[CH4_fluxfinal_aug23_day$R2.CH4 < 0.4,]
```

QC CO2
```{r}
# list of ID names by unique sample measurement (by creating sample-tag)
CO2_fluxfinal_aug23_fix$ID_measurement <- paste(CO2_fluxfinal_aug23_fix$sample, CO2_fluxfinal_aug23_fix$tag, CO2_fluxfinal_aug23_fix$time_of_day, sep = "-")
data_merge_day$ID_measurement <- paste(data_merge_day$sample, data_merge_day$tag, data_merge_day$time_of_day, sep = "-")
target_CO2 <- as.list(CO2_fluxfinal_aug23_fix$ID_measurement)

# function to troubleshoot and visualize low correlation fluxes 
CO2_fix <- function(target_CO2, CO2_fluxfinal_aug23_fix, data_merge_day){
  # select first ID name in list 
  target_ID <- target_CO2[1]
 
  # in fix file, set Time_start and fDOY_end as xlim_lower and xlim_upper 
  working <- CO2_fluxfinal_aug23_fix[CO2_fluxfinal_aug23_fix$ID_measurement == target_ID,]
  records <- data_merge_day[data_merge_day$ID_measurement == target_ID,]
  xlim_lower <- working$Time_start
  xlim_upper <- working$fDOY_end
  ylim_lower <- min(records$CO2_dry) - 5
  ylim_upper <- max(records$CO2_dry) + 5
  
  # plot and label with sample ID name - add to PDF where all are compiled
  ggplot(data_merge_day, aes(x = fDOY, y = CO2_dry))+
    geom_point()+
    theme_classic()+
    xlim(xlim_lower, xlim_upper)+
    ylim(ylim_lower, ylim_upper)+
    ggtitle(target_ID)
}

# test
out_CO2 <- sapply(seq_along(target_CO2), function(x) CO2_fix(target_CO2[[x]], CO2_fluxfinal_aug23_fix, data_merge_day), USE.NAMES= TRUE, simplify = FALSE)

# ggsave(filename = "/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/troubleshooting_fluxes/plotCO2_aug_day.pdf",
#    plot = marrangeGrob(out_CO2, nrow=1, ncol=1),
#    width = 15, height = 9)
```

QC CH4
```{r}
# list of ID names by unique sample measurement (by creating sample-tag)
CH4_fluxfinal_aug23_fix$ID_measurement <- paste(CH4_fluxfinal_aug23_fix$sample, CH4_fluxfinal_aug23_fix$tag, CH4_fluxfinal_aug23_fix$time_of_day, sep = "-")
target_CH4 <- as.list(CH4_fluxfinal_aug23_fix$ID_measurement)

# function to troubleshoot and visualize low correlation fluxes 
CH4_fix <- function(target_CH4, CH4_fluxfinal_aug23_fix, data_merge_day){
  # select first ID name in list 
  target_ID <- target_CH4[1]
 
  # in fix file, set Time_start and fDOY_end as xlim_lower and xlim_upper 
  working <- CH4_fluxfinal_aug23_fix[CH4_fluxfinal_aug23_fix$ID_measurement == target_ID,]
  records <- data_merge_day[data_merge_day$ID_measurement == target_ID,]
  xlim_lower <- working$Time_start
  xlim_upper <- working$fDOY_end
  ylim_lower <- min(records$CH4_dry) - .1
  ylim_upper <- max(records$CH4_dry) + .1
  
  # plot and label with sample ID name - add to PDF where all are compiled
  ggplot(data_merge_day, aes(x = fDOY, y = CH4_dry))+
    geom_point()+
    theme_classic()+
    xlim(xlim_lower, xlim_upper)+
    ylim(ylim_lower, ylim_upper)+
    ggtitle(target_ID)
}

# test
out_CH4 <- sapply(seq_along(target_CH4), function(x) CH4_fix(target_CH4[[x]], CH4_fluxfinal_aug23_fix, data_merge_day), USE.NAMES= TRUE, simplify = FALSE)

# ggsave(filename = "/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/troubleshooting_fluxes/plotCH4_aug_day.pdf",
#    plot = marrangeGrob(out_CH4, nrow=1, ncol=1),
#    width = 15, height = 9)
```

Fix QC flagged fluxes 
```{r}
# display graphs of problem fluxes to define new start/endpoints
problem <- c("DM3-2-afternoon", "DM4-2-midday")

p1 <- ggplot(data_merge_day[data_merge_day$ID_measurement == problem[1],], aes(x = fDOY, y = CH4_dry))+
  geom_point()+
  theme_classic()
ggplotly(p1)

p2 <- ggplot(data_merge_day[data_merge_day$ID_measurement == problem[2],], aes(x = fDOY, y = CH4_dry))+
  geom_point()+
  theme_classic()
ggplotly(p2)

# read in troubleshooting metadata, select for problem fluxes 
trouble <- read.csv("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/troubleshooting_fluxes/troubleshooting_samples.csv")
trouble_test <- trouble[trouble$ID_measurement %in% problem,]

# fix DM3-2-afternoon
fix1 <- data_merge_day[data_merge_day$ID_measurement == problem[1],]
ggplot(fix1, aes(x = fDOY, y = CH4_dry))+
  geom_point()+
  theme_classic()

e <- trouble_test[1,]
new_end <- e$new_end
fixed1 <- fix1[!fix1$fDOY > new_end,]

ggplot(fixed1, aes(x = fDOY, y = CH4_dry))+
  geom_point()+
  theme_classic()

# fix DM4-2-midday
fix2 <- data_merge_day[data_merge_day$ID_measurement == problem[2],]
ggplot(fix2, aes(x = fDOY, y = CH4_dry))+
  geom_point()+
  theme_classic()

s <- trouble_test[2,]
new_start <- s$new_start
fixed2 <- fix2[!fix2$fDOY < new_start,]

ggplot(fixed2, aes(x = fDOY, y = CH4_dry))+
  geom_point()+
  theme_classic()

# now need to calculate fluxes for fixed1 and fixed2 
```

fixed fluxes
```{r}
# create fixed flux df 
fluxes.CH4_aug_fix=data.frame(matrix(NA,nrow=2))
fluxes.CO2_aug_fix=data.frame(matrix(NA,nrow=2))

## Add named columns
# CH4
fluxes.CH4_aug_fix$Time_start=0
fluxes.CH4_aug_fix$flux.CH4=0
fluxes.CH4_aug_fix$R2.CH4=0
fluxes.CH4_aug_fix$p.CH4=0

# CO2
fluxes.CO2_aug_fix$Time_start=0
fluxes.CO2_aug_fix$flux.CO2=0
fluxes.CO2_aug_fix$R2.CO2=0
fluxes.CO2_aug_fix$p.CO2=0

## Remove initial empty column
fluxes.CO2_aug_fix=fluxes.CO2_aug_fix[,-1]
fluxes.CH4_aug_fix=fluxes.CH4_aug_fix[,-1]

# CH4
# fixed1: DM3-2-afternoon
# linear model and statistics
mod <- lm(CH4_dry_umol~fDOY, data = fixed1)
# Save flux rate and R2 and p-value of slope in corresponding row of dataframe
fluxes.CH4_aug_fix[1,]$Time_start <- as.numeric(min(fixed1$fDOY))
fluxes.CH4_aug_fix[1,]$flux.CH4=unique(coef(mod)[2]/fixed1$chamber_SA_m2) # chamber surface area
fluxes.CH4_aug_fix[1,]$R2.CH4=summary(mod)$r.squared
fluxes.CH4_aug_fix[1,]$p.CH4=summary(mod)$coefficients[2,4]

# fixed2: DM4-2-midday
mod <- lm(CH4_dry_umol~fDOY, data = fixed2)
# Save flux rate and R2 and p-value of slope in corresponding row of dataframe
fluxes.CH4_aug_fix[2,]$Time_start <- as.numeric(min(fixed2$fDOY))
fluxes.CH4_aug_fix[2,]$flux.CH4=unique(coef(mod)[2]/fixed2$chamber_SA_m2) # chamber surface area
fluxes.CH4_aug_fix[2,]$R2.CH4=summary(mod)$r.squared
fluxes.CH4_aug_fix[2,]$p.CH4=summary(mod)$coefficients[2,4]

# CO2
# fixed1: DM3-2-afternoon
# linear model and statistics
mod <- lm(CO2_dry_mmol~fDOY, data = fixed1)
# Save flux rate and R2 and p-value of slope in corresponding row of dataframe
fluxes.CO2_aug_fix[1,]$Time_start <- as.numeric(min(fixed1$fDOY))
fluxes.CO2_aug_fix[1,]$flux.CO2=unique(coef(mod)[2]/fixed1$chamber_SA_m2) # chamber surface area
fluxes.CO2_aug_fix[1,]$R2.CO2=summary(mod)$r.squared
fluxes.CO2_aug_fix[1,]$p.CO2=summary(mod)$coefficients[2,4]

# fixed2: DM4-2-midday
mod <- lm(CO2_dry_mmol~fDOY, data = fixed2)
# Save flux rate and R2 and p-value of slope in corresponding row of dataframe
fluxes.CO2_aug_fix[2,]$Time_start <- as.numeric(min(fixed2$fDOY))
fluxes.CO2_aug_fix[2,]$flux.CO2=unique(coef(mod)[2]/fixed2$chamber_SA_m2) # chamber surface area
fluxes.CO2_aug_fix[2,]$R2.CO2=summary(mod)$r.squared
fluxes.CO2_aug_fix[2,]$p.CO2=summary(mod)$coefficients[2,4]

```

add fixed fluxes to final df 
```{r}
# first pull out incorrect fluxes from original df 
a <- data_merge_day[data_merge_day$ID_measurement %in% problem,]
original_start_fix <- aggregate(fDOY ~ sample, a, function(x) min(x))
original_start_fix <-  original_start_fix[['fDOY']]

# add in corrected fluxes - ch4
working <- CH4_fluxfinal_aug23_day[CH4_fluxfinal_aug23_day$Time_start %in% original_start_fix,]
# remove so that corrected version can be merged back in 
CH4_fluxfinal_aug23_day <- CH4_fluxfinal_aug23_day[!CH4_fluxfinal_aug23_day$Time_start %in% original_start_fix,]

working[2,]$flux.CH4 <- fluxes.CH4_aug_fix[1,]$flux.CH4
working[2,]$R2.CH4 <- fluxes.CH4_aug_fix[1,]$R2.CH4
working[2,]$p.CH4 <- fluxes.CH4_aug_fix[1,]$p.CH4

working[1,]$flux.CH4 <- fluxes.CH4_aug_fix[2,]$flux.CH4
working[1,]$R2.CH4 <- fluxes.CH4_aug_fix[2,]$R2.CH4
working[1,]$p.CH4 <- fluxes.CH4_aug_fix[2,]$p.CH4

# add in corrected fluxes - co2
working_c <- CO2_fluxfinal_aug23_day[CO2_fluxfinal_aug23_day$Time_start %in% original_start_fix,]
# remove so that corrected version can be merged back in 
CO2_fluxfinal_aug23_day <- CO2_fluxfinal_aug23_day[!CO2_fluxfinal_aug23_day$Time_start %in% original_start_fix,]

working_c[1,]$flux.CO2 <- fluxes.CO2_aug_fix[1,]$flux.CO2
working_c[1,]$R2.CO2 <- fluxes.CO2_aug_fix[1,]$R2.CO2
working_c[1,]$p.CO2 <- fluxes.CO2_aug_fix[1,]$p.CO2

working_c[2,]$flux.CO2 <- fluxes.CO2_aug_fix[2,]$flux.CO2
working_c[2,]$R2.CO2 <- fluxes.CO2_aug_fix[2,]$R2.CO2
working_c[2,]$p.CO2 <- fluxes.CO2_aug_fix[2,]$p.CO2

# bring back together with finalized, fixed fluxes 
CH4_fluxfinal_aug23_day_fixed <- rbind(CH4_fluxfinal_aug23_day, working)
CO2_fluxfinal_aug23_day_fixed <- rbind(CO2_fluxfinal_aug23_day, working_c)

# quick recheck on QC (anything that is <0.4 is due to near-zero fluxes)
# co2 
mean(CO2_fluxfinal_aug23_day_fixed$R2.CO2)
ggplot(CO2_fluxfinal_aug23_day_fixed, aes(R2.CO2))+
  geom_histogram()+
  theme_classic()

# ch4 
mean(CH4_fluxfinal_aug23_day_fixed$R2.CH4)
ggplot(CH4_fluxfinal_aug23_day_fixed, aes(R2.CH4))+
  geom_histogram()+
  theme_classic()
```

set non-significant fluxes to 0 
```{r}
# define ID_measurement column 
CO2_fluxfinal_aug23_day_fixed$ID_measurement <- paste(CO2_fluxfinal_aug23_day_fixed$sample, CO2_fluxfinal_aug23_day_fixed$tag, CO2_fluxfinal_aug23_day_fixed$time_of_day, sep = "-")
CH4_fluxfinal_aug23_day_fixed$ID_measurement <- paste(CH4_fluxfinal_aug23_day_fixed$sample, CH4_fluxfinal_aug23_day_fixed$tag, CH4_fluxfinal_aug23_day_fixed$time_of_day, sep = "-")

# pull out samples that have p < 0.05
table(CO2_fluxfinal_aug23_day_fixed$p.CO2<0.05) # all CO2 is fine (either removed or fluxes are fixed)
table(CH4_fluxfinal_aug23_day_fixed$p.CH4<0.05)
a <- CH4_fluxfinal_aug23_day_fixed[CH4_fluxfinal_aug23_day_fixed$p.CH4>0.05,]

# for CH4
CH4_to_zero <- a$ID_measurement
assign_zeros_CH4 <- c('flux.CH4', 'R2.CH4', 'p.CH4')

# function to reassign nonsignificant fluxes to 0 
zero_flux <- function(CH4_fluxfinal_aug23_day_fixed, CH4_to_zero, assign_zeros_CH4) {
  # Replace columns in assign_zeros_CH4 with 0 for ID_measurement values in CH4_to_zero
  CH4_fluxfinal_aug23_day_fixed <- CH4_fluxfinal_aug23_day_fixed %>%
    mutate_at(vars(all_of(assign_zeros_CH4)), ~ ifelse(ID_measurement %in% CH4_to_zero, 0, .))
  
  return(CH4_fluxfinal_aug23_day_fixed)
}

CH4_fluxfinal_aug23_day_fixed_zeroadj <- zero_flux(CH4_fluxfinal_aug23_day_fixed, CH4_to_zero, assign_zeros_CH4)
```

EXPORT
```{r}
## Export fluxes as .csv file
# CO2 (units are mmol/day)
# write.csv(CO2_fluxfinal_aug23_day_fixed,"/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/finalfluxes/CO2_fluxfinal_aug23_day_fixed.csv", row.names = FALSE)

# CH4 (units are umol/day)
# write.csv(CH4_fluxfinal_aug23_day_fixed_zeroadj,"/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/finalfluxes/CH4_fluxfinal_aug23_day_fixed.csv", row.names = FALSE)
```

calculate Q10
```{r}
# first look - isolate a single mound 
DM <- CH4_day[CH4_day$sample == "DM1",]
plot(DM$avg_respT, DM$flux.CH4)

# look at range of temperatures 
min(DM$avg_respT); max(DM$avg_respT)

test <- DM[order(DM$avg_respT),]
t <- test$avg_resp
r <- test$flux.CH4
Q10(R_vec = r, T_vec = t)

# get list of mound IDs 
x <- unique(CH4_day$sample)

# Q10 associated with each individual mound
for (i in x){
  df <- CH4_day[CH4_day$sample == i,]
  t <- df$avg_resp
  r <- df$flux.CH4
  print(Q10(R_vec = r, T_vec = t))
}

# flux ~ temperature relationships
par(mfrow = c(3, 3))
for(i in x){
  df <- CH4_day[CH4_day$sample == i,]
  plot(df$avg_respT, df$flux.CH4, main = paste("Mound", i),)
}

# TODO mound as a random effect in Q10 model, also try model with mound averages
# same thing with CO2 
```
