==> scripts/6.NT_process.Rmd <==
---
title: "9.naked_termite"
author: "abbey yatsko"
date: "2024-03-15"
output: html_document
---

workspace 
```{r}
setwd("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/NT/LGR_files")
filenames <- list.files(pattern='f000',full.names=T)

library(data.table)
library(ggplot2)
library(ggpubr)
library(sqldf)
library(dplyr)
library(tidyverse)
library(stringr)
library(plyr)
```

data process 
```{r}
# read in LGR datafiles to list and go from .txt to .csv 
dat <- lapply(filenames,read.csv,skip=1)
dat <- rbindlist(dat)
dat <- subset(dat, X.CH4._ppm!="NA")
as.character(dat$Time)

date_time <- strptime(dat$Time,format='%d/%m/%Y %H:%M:%S')

# add year,month,day,JD,hour,min,sec columns to dataframe
Year <- as.numeric(format(date_time,'%Y'))
Month <- as.numeric(format(date_time,'%m'))
Day <- as.numeric(format(date_time,'%d'))
fDOY <- as.numeric(julian(date_time,'2024-01-01'))  #Change for year
Hour <- as.numeric(format(date_time,'%k'))
Min <- as.numeric(format(date_time,'%M'))
Sec <- as.numeric(format(date_time,'%S'))
data_clean <- cbind(date_time,Year,Month,Day,fDOY,Hour,Min,Sec,dat)

# save LGR data as dt
data_clean <- data.table(data_clean)
ggplot(data_clean, aes(date_time, X.CH4.d_ppm)) +
  geom_point() +
  theme(axis.text.x = element_text(angle = 90, hjust = 1)) +
  ggtitle("flux")+
  scale_x_datetime(limits = c(as.POSIXct("2024-03-01 11:58:00"), as.POSIXct("2024-03-01 13:19:50")))

# plotly_code <- ggplotly(samp)
```

merge with metadata
```{r}
meta_TMI <- read.csv('/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/NT/arduino_files/NTmetadata.csv')
as.Date(meta_TMI$measurement_day, '%m/%d/%Y')

# define start time
date_time_start <- strptime(paste(meta_TMI$measurement_day,meta_TMI$flux_start),'%m/%d/%Y %H:%M:%S')

## Add 30 seconds to set buffered start time
date_time_start <- date_time_start+60

## Add 90 sec to determine end time
date_time_end <- date_time_start+180

## Add fDOY columns for start and endtime to dataframe
fDOY_start <- as.numeric(julian(date_time_start,'2024-01-01'))  #Change for year
fDOY_end <- as.numeric(julian(date_time_end,'2024-01-01'))  #Change for year
meta_TMI <- cbind(fDOY_start,fDOY_end,meta_TMI)

# change sample to factor level variables 
meta_TMI$ID <- as.factor(meta_TMI$ID) 

# rename co2 and ch4 columns because naming format including '.' messes up the sql code 
data_clean <- data_clean %>%
  dplyr::rename(CH4 = X.CH4._ppm,
                CO2 = X.CO2._ppm,
                CH4_dry = X.CH4.d_ppm,
                CO2_dry = X.CO2.d_ppm)
```

merge LGR files with metadata 
```{r}
# complete merge for respiration data and metadata (aug)
data_merge <- sqldf("select data_clean.fDOY, data_clean.CO2_dry, data_clean.CH4_dry, meta_TMI.ID, meta_TMI.material, meta_TMI.chunk, meta_TMI.avg_temp_C, meta_TMI.collar_volume_cm3,  meta_TMI.mass_mound_material_g, meta_TMI.all_mass_g, meta_TMI.mass_termite_g, meta_TMI.material_volume_cm3, meta_TMI.fDOY_start 
from data_clean LEFT JOIN meta_TMI ON (data_clean.fDOY BETWEEN meta_TMI.fDOY_start AND meta_TMI.fDOY_end)")

# use 'complete.cases'
data_merge <- data_merge[complete.cases(data_merge), ]

# check to make sure all of the samples are there (52 total)
summary(unique(data_merge$ID))
```

Visually check clipped bits of data 
```{r}
# co2 
par(mfrow=c(2,1),mar=c(4,4,1,1))
with(data_clean,plot(fDOY,CO2_dry,ylim=c(400,900), xlim=c(65.42015, 65.6151)))
with(data_merge,plot(fDOY,CO2_dry,ylim=c(400,900), xlim=c(65.42015, 65.6151)))

# ch4 
par(mfrow=c(2,1),mar=c(4,4,1,1))
with(data_clean,plot(fDOY,CH4_dry,ylim=c(1.8, 2.5), xlim=c(65.42015, 65.6151)))
with(data_merge,plot(fDOY,CH4_dry,ylim=c(1.8,2.5), xlim=c(65.42015, 65.6151)))
```

convert to moles
```{r}
# chamber volumes are in cm3 - convert to m3 for the next step 
data_merge$collar_volume_m3 <- data_merge$collar_volume_cm3*0.000001
data_merge$material_volume_m3 <- as.numeric(data_merge$material_volume_cm3)*0.000001
data_merge$volume_m3 <- data_merge$collar_volume_m3-data_merge$material_volume_m3

# CH4
data_merge$CH4_dry_L=
  # parts CH4 per million parts air * volume of air in chamber (m3) * 1000 L per m3
  (data_merge$CH4_dry/1000000) * data_merge$volume_m3 * 1000 

# CO2
data_merge$CO2_dry_L=
  # parts CO2 per million parts air * volume of air in chamber (m3) * 1000 L per m3
  (data_merge$CO2_dry/1000000) * data_merge$volume_m3 * 1000 

# Use ideal gas law to calculate umol of CH4 or mmol of CO2
# CH4
data_merge$CH4_dry_umol=
  # (atm pressure * L CH4) / (R in L*atm/?K*mol * ?K temp) * 10^6 umol/mol
  ((1*data_merge$CH4_dry_L)/(0.08206*(data_merge$avg_temp_C+273)))*10^6

# CO2
data_merge$CO2_dry_mmol=
  # (atm pressure * L CO2) / (R in L*atm/?K*mol * ?K temp) * 10^3 mmol/mol
  ((1*data_merge$CO2_dry_L)/(0.08206*(data_merge$avg_temp_C+273)))*10^3
```

Calculate chamber fluxes
```{r}
## Identify start of fluxes
Time_start <- meta_TMI[,c("fDOY_start","ID")]
flux.times=unique(Time_start$fDOY_start)

# Create new dataframes to hold final fluxes 
fluxes.CH4=data.frame(matrix(NA,nrow=length(flux.times)))
fluxes.CO2=data.frame(matrix(NA,nrow=length(flux.times)))

## Add named columns
# CH4
fluxes.CH4$Time_start=flux.times
fluxes.CH4$flux.CH4=0
fluxes.CH4$R2.CH4=0
fluxes.CH4$p.CH4=0
fluxes.CH4$sd.CH4=0

# CO2
fluxes.CO2$Time_start=flux.times
fluxes.CO2$flux.CO2=0
fluxes.CO2$R2.CO2=0
fluxes.CO2$p.CO2=0
fluxes.CO2$sd.CO2=0

## Remove initial empty column
fluxes.CO2=fluxes.CO2[,-1]
fluxes.CH4=fluxes.CH4[,-1]

#i = flux.times[2]
## For each start time (aug 2023)
for (i in flux.times) {
  ## CH4 ##
  # Subset data for one chamber measurement
  temp1=subset(data_merge,fDOY_start==i)
  # Set corresponding row of output table
  j=which(flux.times==i)
  # Determine if start time has a CH4 flux
  if (nrow(temp1)>0) {
    # If so:  
    # Calulate flux in umol/day using linear regression
    mod=with(temp1,lm(CH4_dry_umol~fDOY))
    # Save flux rate and R2 and p-value of slope in corresponding row of dataframe
    # flux rate
    fluxes.CH4$flux.CH4[j]=coef(mod)[2]
    # R2 of slope
    fluxes.CH4$R2.CH4[j]=summary(mod)$r.squared
    # p-value of slope
    fluxes.CH4$p.CH4[j]=summary(mod)$coefficients[2,4]
    # sd 
    fluxes.CH4$sd.CH4[j] <- summary(mod)$coefficients["fDOY", "Std. Error"]

    # If not:
    # Fill rows of table with NA    
  } else {
    fluxes.CH4$flux.CH4[j]=NA
    fluxes.CH4$R2.CH4[j]=NA
    fluxes.CH4$p.CH4[j]=NA
    fluxes.CH4$sd.CH4[j]=NA
  }
  ## CO2 ##
  # Subset data for one chamber measurement
  temp2=subset(data_merge,fDOY_start==i)
  # Calulate flux in mmol/day using linear regression
  mod=with(temp2,lm(CO2_dry_mmol~fDOY))
  # Save flux rate and R2 and p-value of slope in corresponding row of dataframe
  # flux rate
  fluxes.CO2$flux.CO2[j]=coef(mod)[2]
  # R2 of slope
  fluxes.CO2$R2.CO2[j]=summary(mod)$r.squared
  # p-value of slope
  fluxes.CO2$p.CO2[j]=summary(mod)$coefficients[2,4]
  # sd 
  fluxes.CO2$sd.CO2[j] <- summary(mod)$coefficients["fDOY", "Std. Error"]
}
```

finalize and bring back in metadata 
```{r}
meta_TMI <- meta_TMI %>%
  dplyr::rename(Time_start = fDOY_start)
CO2_fluxfinal <- merge(fluxes.CO2, meta_TMI, by="Time_start")
CH4_fluxfinal <- merge(fluxes.CH4, meta_TMI, by="Time_start")

# generate summary table 
CH4_summary <- CH4_fluxfinal[, c("chunk","species", "ID", "material", "all_mass_g", "mass_mound_material_g", "mass_termite_g", "material_volume_cm3", "flux.CH4", "sd.CH4", "R2.CH4", "p.CH4", "Time_start", "fDOY_end")]
CO2_summary <- CO2_fluxfinal[, c("ID", "flux.CO2", "sd.CO2", "R2.CO2", "p.CO2")]
flux_summary <- left_join(CH4_summary, CO2_summary, by = "ID")

# define termite emission factors (TEF), mound emission factor (MEF), termite respiration factor (TRF), mound respiration factor (MRF)
flux_summary <- flux_summary %>%
  group_by(chunk) %>%
  mutate(TEF_umol_g_d = ifelse(material == "T", flux.CH4 / mass_termite_g, NA), 
         MEF_umol_g_d = ifelse(material == "M", flux.CH4 / mass_mound_material_g, NA), 
         AEF_umol_g_d = ifelse(material == "A", flux.CH4 / all_mass_g, NA), 
         TRF_mmol_g_d = ifelse(material == "T", flux.CO2 / mass_termite_g, NA), 
         MRF_mmol_g_d = ifelse(material == "M", flux.CO2 / mass_mound_material_g, NA), 
         ARF_mmol_g_d = ifelse(material == "A", flux.CO2 / all_mass_g, NA))

# clean out unusable samples 
# remove first two coptotermes mounds NT07-09, NT10-12, NT0811TERM, 
rm <- c("NT0811TERM", "NT07", "NT08", "NT09", "NT10", "NT11", "NT12")
flux_summary_f <- subset(flux_summary, !ID %in% rm)
```

preliminary viz 
```{r}
# emission factor
ggplot(flux_summary_f, aes(x = species, y = TEF_umol_g_d, fill = species))+
  geom_boxplot()+
  geom_jitter()+
  theme_classic()+
  theme(legend.position = "none")

# flux by material 
meth <- ggplot(flux_summary_f, aes(x = species, y = flux.CH4, fill = material))+
  geom_boxplot()+
  geom_jitter()+
  theme_classic()+
  xlab("")

co2 <- ggplot(flux_summary_f, aes(x = species, y = flux.CO2, fill = material))+
  geom_boxplot()+
  geom_jitter()+
  theme_classic()+
  xlab("")

ggarrange(meth, co2, common.legend = TRUE)
```

QC check for all measurements
```{r}
# check r2 values 
# co2 
mean(flux_summary_f$R2.CO2)
ggplot(flux_summary_f, aes(R2.CO2))+
  geom_histogram()+
  theme_classic()

# ch4 
mean(flux_summary_f$R2.CH4)
ggplot(flux_summary_f, aes(R2.CH4))+
  geom_histogram()+
  theme_classic()

# check p values
table(flux_summary_f$p.CO2<0.05)
table(flux_summary_f$p.CH4<0.05)

# extract everything with R2 < 0.4
CO2_fluxfinal_NT_fix <- flux_summary_f[flux_summary_f$R2.CO2 < 0.4,]
CH4_fluxfinal_NT_fix <- flux_summary_f[flux_summary_f$R2.CH4 < 0.4,]
```

QC CO2 
```{r}
# list of ID names by unique sample measurement (by creating sample-tag)
CO2_fluxfinal_NT_fix$ID_measurement <- paste(CO2_fluxfinal_NT_fix$ID, CO2_fluxfinal_NT_fix$material, sep = "-")
data_merge$ID_measurement <- paste(data_merge$ID, data_merge$material, sep = "-")
target_CO2 <- as.list(CO2_fluxfinal_NT_fix$ID_measurement)

# function to troubleshoot and visualize low correlation fluxes 
CO2_fix <- function(target_CO2, CO2_fluxfinal_NT_fix, data_merge){
  # select first ID name in list 
  target_ID <- target_CO2[1]
 
  # in fix file, set Time_start and fDOY_end as xlim_lower and xlim_upper 
  working <- CO2_fluxfinal_NT_fix[CO2_fluxfinal_NT_fix$ID_measurement == target_ID,]
  records <- data_merge[data_merge$ID_measurement == target_ID,]
  xlim_lower <- working$Time_start  
  xlim_upper <- working$fDOY_end   
  ylim_lower <- min(records$CO2_dry) - 5
  ylim_upper <- max(records$CO2_dry) + 5
  
  # plot and label with sample ID name - add to PDF where all are compiled
  ggplot(data_merge, aes(x = fDOY, y = CO2_dry))+
    geom_point()+
    theme_classic()+
    xlim(xlim_lower, xlim_upper)+
    ylim(ylim_lower, ylim_upper)+
    ggtitle(target_ID)
}

# test
out_CO2 <- sapply(seq_along(target_CO2), function(x) CO2_fix(target_CO2[[x]], CO2_fluxfinal_NT_fix, data_merge), USE.NAMES= TRUE, simplify = FALSE)

# ggsave(filename = "/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/troubleshooting_fluxes/plotCO2_NT.pdf",
#    plot = marrangeGrob(out_CO2, nrow=1, ncol=1),
#    width = 15, height = 9)
```

QC CH4
```{r}
# list of ID names by unique sample measurement (by creating sample-tag)
CH4_fluxfinal_NT_fix$ID_measurement <- paste(CH4_fluxfinal_NT_fix$ID, CH4_fluxfinal_NT_fix$material, sep = "-")
data_merge$ID_measurement <- paste(data_merge$ID, data_merge$material, sep = "-")
target_CH4 <- as.list(CH4_fluxfinal_NT_fix$ID_measurement)

# function to troubleshoot and visualize low correlation fluxes 
CH4_fix <- function(target_CH4, CH4_fluxfinal_NT_fix, data_merge){
  # select first ID name in list 
  target_ID <- target_CH4[1]
 
  # in fix file, set Time_start and fDOY_end as xlim_lower and xlim_upper 
  working <- CH4_fluxfinal_NT_fix[CH4_fluxfinal_NT_fix$ID_measurement == target_ID,]
  records <- data_merge[data_merge$ID_measurement == target_ID,]
  xlim_lower <- working$Time_start
  xlim_upper <- working$fDOY_end
  ylim_lower <- min(records$CH4_dry) - .1
  ylim_upper <- max(records$CH4_dry) + .1
  
  # plot and label with sample ID name - add to PDF where all are compiled
  ggplot(data_merge, aes(x = fDOY, y = CH4_dry))+
    geom_point()+
    theme_classic()+
    xlim(xlim_lower, xlim_upper)+
    ylim(ylim_lower, ylim_upper)+
    ggtitle(target_ID)
}

# test
out_CH4 <- sapply(seq_along(target_CH4), function(x) CH4_fix(target_CH4[[x]], CH4_fluxfinal_NT_fix, data_merge), USE.NAMES= TRUE, simplify = FALSE)

# ggsave(filename = "/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/troubleshooting_fluxes/plotCH4_NT.pdf",
#    plot = marrangeGrob(out_CH4, nrow=1, ncol=1),
#    width = 15, height = 9)
```

no fix, set non-significant fluxes to 0
```{r}
# apply to main df 
flux_summary_f$ID_measurement <- paste(flux_summary_f$ID, flux_summary_f$material, sep = "-")

# pull out samples that have p < 0.05
table(flux_summary_f$p.CO2<0.05) # all CO2 is fine (either removed or fluxes are fixed)
table(flux_summary_f$p.CH4<0.05)
a <- flux_summary_f[flux_summary_f$p.CH4>0.05,]
CH4_to_zero <- a$ID_measurement
assign_zeros_CH4 <- c('flux.CH4', 'sd.CH4', 'R2.CH4', 'p.CH4')

# function to reassign nonsignificant fluxes to 0 
zero_flux <- function(flux_summary_f, CH4_to_zero, assign_zeros_CH4) {
  # Replace columns in assign_zeros_CH4 with 0 for ID_measurement values in CH4_to_zero
  flux_summary_f <- flux_summary_f %>%
    mutate_at(vars(all_of(assign_zeros_CH4)), ~ ifelse(ID_measurement %in% CH4_to_zero, 0, .))
  
  return(flux_summary_f)
}

flux_summary_f_zeroadj <- zero_flux(flux_summary_f, CH4_to_zero, assign_zeros_CH4)
```

unit conversion and create proportional flux variables 
```{r}
# rename EF columns with correct units for comparison: go from umol CH4 g-1 d-1 --> ug CH4 g-1 h-1
flux_summary_f_zeroadj$TEF_ug_g_h <- (flux_summary_f_zeroadj$TEF_umol_g_d*16.04)/24

# proportional fluxes df
p_fluxes <- flux_summary_f_zeroadj[c("chunk", "species", "mass_termite_g", "material_volume_cm3", "material", "TEF_umol_g_d", "TEF_ug_g_h", "flux.CH4", "flux.CO2")]
p_fluxes_f <- p_fluxes %>%
  mutate(flux.CH4_A = ifelse(material == "A", flux.CH4, NA),
         flux.CH4_T = ifelse(material == "T", flux.CH4, NA),
         flux.CH4_M = ifelse(material == "M", flux.CH4, NA), 
         flux.CO2_A = ifelse(material == "A", flux.CO2, NA),
         flux.CO2_T = ifelse(material == "T", flux.CO2, NA),
         flux.CO2_M = ifelse(material == "M", flux.CO2, NA), 
         mass_termite_g = ifelse(mass_termite_g > 0, mass_termite_g, NA), 
         TEF_umol_g_d = ifelse(TEF_umol_g_d > 0, TEF_umol_g_d, NA),
         TEF_ug_g_h = ifelse(TEF_ug_g_h > 0, TEF_ug_g_h, NA),
         material_volume_cm3 = ifelse(material_volume_cm3 > 0, material_volume_cm3, NA)) %>%
  select(-material, -flux.CH4, -flux.CO2)%>%
  group_by(chunk) %>%
  filter_at(vars(-group_cols()), any_vars(!is.na(.))) %>%
  summarise_all(~first(.[!is.na(.)]))

# methane flux proportions (out of termite + microbe separate)
p_fluxes_f$prop_T_CH4 <- p_fluxes_f$flux.CH4_T/(p_fluxes_f$flux.CH4_T+p_fluxes_f$flux.CH4_M)
p_fluxes_f$prop_M_CH4 <- p_fluxes_f$flux.CH4_M/(p_fluxes_f$flux.CH4_T+p_fluxes_f$flux.CH4_M)

# denominator is undisturbed termite + microbe (does less well in summing to 1)
p_fluxes_f$prop_T_CH4_A <- p_fluxes_f$flux.CH4_T/(p_fluxes_f$flux.CH4_A)
p_fluxes_f$prop_M_CH4_A <- p_fluxes_f$flux.CH4_M/(p_fluxes_f$flux.CH4_A)

# carbon dioxide flux proportions (out of termite + microbe separate)
p_fluxes_f$prop_T_CO2 <- p_fluxes_f$flux.CO2_T/(p_fluxes_f$flux.CO2_T+p_fluxes_f$flux.CO2_M)
p_fluxes_f$prop_M_CO2 <- p_fluxes_f$flux.CO2_M/(p_fluxes_f$flux.CO2_T+p_fluxes_f$flux.CO2_M)

# denominator is undisturbed termite + microbe (does less well in summing to 1)
p_fluxes_f$prop_T_CO2_A <- p_fluxes_f$flux.CO2_T/(p_fluxes_f$flux.CO2_A)
p_fluxes_f$prop_M_CO2_A <- p_fluxes_f$flux.CO2_M/(p_fluxes_f$flux.CO2_A)
```

EXPORT
```{r}
## Export fluxes as .csv file
# write.csv(flux_summary_f_zeroadj, "/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/finalfluxes/all_fluxes_NT.csv", row.names = FALSE)

# write.csv(p_fluxes_f, "/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/finalfluxes/p_fluxes_f.csv", row.names = FALSE)
```
