==> scripts/7.flux_compile.Rmd <==
---
title: "6. flux_compile"
author: "abbey yatsko"
date: "2024-03-15"
output: html_document
---
libraries
```{r}
library(tidyverse)
library(plyr)
```

read in data
```{r data load}
may_CO2 <- read.csv("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/finalfluxes/CO2_fluxfinal_may22_fixed.csv")
may_CH4 <- read.csv("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/finalfluxes/CH4_fluxfinal_may22_fixed.csv")

nov_CO2 <- read.csv("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/finalfluxes/CO2_fluxfinal_nov22_fixed.csv")
nov_CH4 <- read.csv("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/finalfluxes/CH4_fluxfinal_nov22_fixed.csv")

aug_CO2 <- read.csv("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/finalfluxes/CO2_fluxfinal_aug23_fixed.csv")
aug_CH4 <- read.csv("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/finalfluxes/CH4_fluxfinal_aug23_fixed.csv")

feb_CO2 <- read.csv("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/finalfluxes/CO2_fluxfinal_feb24_fixed.csv")
feb_CH4 <- read.csv("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/finalfluxes/CH4_fluxfinal_feb24_fixed.csv")

CH4_daily <- read.csv("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/finalfluxes/CH4_fluxfinal_aug23_day_fixed.csv")
CO2_daily <- read.csv("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/finalfluxes/CO2_fluxfinal_aug23_day_fixed.csv")

precip <- read.csv("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/supp_data/precip.csv")
mound_SA <- read.csv("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/supp_data/mound_SA.csv")
```

data clean and reorganize
```{r}
# lists of CO2 and CH4 files from different seasons
CH4_all <- list(may_CH4, nov_CH4, aug_CH4, feb_CH4)
CO2_all <- list(may_CO2, nov_CO2, aug_CO2, feb_CO2)

# create new column for methane:co2
# for each element in list 
x <- data.frame(CO2_all[1])
y <- data.frame(CH4_all[1])

full_df <- function(x, y){
  # merge CO2 and CH4 df together 
  all <- full_join(x, y)

  # convert CO2 from mmol to umol (in order to match CH4 units, conversion factor 1000umol = 1mmol)
  all$flux.CO2_umol <- all$flux.CO2*1000
  all <- all %>% relocate(flux.CO2_umol, .after = flux.CO2)
  all <- all %>% relocate(flux.CH4, .after = flux.CO2_umol)
  all <- all %>% relocate(R2.CH4, .after = R2.CO2)
  all <- all %>% relocate(p.CH4, .after = p.CO2)
  
  # define column for ratio of CH4:CO2
  all$CH4_CO2 <-  all$flux.CH4/all$flux.CO2_umol
  all <- all %>% relocate(CH4_CO2, .after = flux.CH4)

}

out <- lapply(seq_along(CH4_all), function(x) full_df(CH4_all[[x]], CO2_all[[x]]))
all_fluxes <- do.call("rbind.fill", out)

# rename species with scientific names 
all_fluxes$species_s <- ifelse(all_fluxes$species == "Ami", "A. laurensis","")
all_fluxes$species_s <- ifelse(all_fluxes$species == "Nasuti", "N. magnus", all_fluxes$species_s)
all_fluxes$species_s <- ifelse(all_fluxes$species == "Copto", "C. acinaciformis", all_fluxes$species_s)

# create unique ID for mounds in different sampling periods
all_fluxes$ID_mound <- paste(all_fluxes$sample, all_fluxes$campaign, sep = "-")

# filter out measurements without species recorded
all_fluxes <- all_fluxes %>%
  filter(species_s != "")

# add in precipitation data 
# add three columns: precip_1mo, precip_2mo, precip_3mo
all_fluxes$precip_1mo <-  ifelse(all_fluxes$campaign == "may22", precip$may_1mo_sum,
                                ifelse(all_fluxes$campaign == "aug23", precip$aug_1mo_sum,
                                       ifelse(all_fluxes$campaign == "nov22", precip$nov_1mo_sum,
                                              ifelse(all_fluxes$campaign == "feb24", precip$feb_1mo_sum, NA))))
all_fluxes$precip_2mo <-  ifelse(all_fluxes$campaign == "may22", precip$may_2mo_sum,
                                ifelse(all_fluxes$campaign == "aug23", precip$aug_2mo_sum,
                                       ifelse(all_fluxes$campaign == "nov22", precip$nov_2mo_sum,
                                              ifelse(all_fluxes$campaign == "feb24", precip$feb_2mo_sum, NA))))
all_fluxes$precip_3mo <- ifelse(all_fluxes$campaign == "may22", precip$may_3mo_sum,
                                ifelse(all_fluxes$campaign == "aug23", precip$aug_3mo_sum,
                                       ifelse(all_fluxes$campaign == "nov22", precip$nov_3mo_sum,
                                              ifelse(all_fluxes$campaign == "feb24", precip$feb_3mo_sum, NA))))

# EXPORT - all fluxes
# write.csv(all_fluxes, "/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/finalfluxes/all_fluxes.csv", row.names = FALSE)

# EXPORT - mound-only fluxes
all_fluxes_mound <- all_fluxes[all_fluxes$flux_source =="m",]
# write.csv(all_fluxes_mound, "/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/finalfluxes/all_fluxes_mound.csv", row.names = FALSE)

# EXPORT - soil-only fluxes
all_fluxes_soil <- all_fluxes[all_fluxes$flux_source =="s",]
# write.csv(all_fluxes_soil, "/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/finalfluxes/all_fluxes_soil.csv", row.names = FALSE)
```

data summary from compiled list 
```{r}
# reorder levels of campaigns 
all_fluxes$campaign <- factor(all_fluxes$campaign, levels = c("may22", "aug23", "nov22", "feb24"))

# summary of how mounds were measured for each campaign
detach(package:plyr)
summary_counts <- all_fluxes %>%
  group_by(campaign, species_s) %>%
    summarise(unique_samples = n_distinct(sample))

# measurements per mound 
measurements_per_mound <- aggregate(all_fluxes$tag , by=list(all_fluxes$ID_mound ), FUN=length)

measurements_per_mound <- measurements_per_mound %>%
  rename(ID_mound = Group.1,
         count = x)

# how many measurements were made for each mound in the resampling campaign (should be 4 or 5)
targets1 <- unique(nov_CH4$sample)
targets2 <- unique(aug_CH4$sample)
targets3 <- unique(feb_CH4$sample)
targets <- union(targets1, targets2)
targets_fin <- union(targets, targets3)

all_fluxes_resample <- all_fluxes[all_fluxes$sample %in% targets_fin, ]

# measurements across all 4 campaigns 
resample_repeats <- all_fluxes_resample %>%
  group_by(sample) %>%
    summarise(repeats = n_distinct(campaign), 
              species = unique(species_s))

# filter out mounds with fewer than 4 time points 
seasonal_list <- resample_repeats[resample_repeats$repeats>3,]
seasonal_list <- seasonal_list$sample

all_fluxes_resample_reps <- all_fluxes_resample[all_fluxes_resample$sample %in% seasonal_list,]

# EXPORT - seasonal resampled fluxes
# write.csv(all_fluxes_resample_reps, "/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/finalfluxes/all_fluxes_resample_reps.csv", row.names = FALSE)

# working df for coarse mound averages 
all_fluxes_mound_avg <- all_fluxes_mound %>% group_by(ID_mound, sample, species_s) %>% 
  summarise(mean_mound_fluxCH4=mean(flux.CH4),
            mean_mound_fluxCO2=mean(flux.CO2_umol), 
            mean_mound_fluxCH4_CO2=mean(CH4_CO2), 
            campaign = campaign, 
            wall_thickness_mm = mean(wall_thickness_mm), 
            temp = mean(avg_respT), 
            precip_1mo = precip_1mo, 
            precip_2mo = precip_2mo, 
            precip_3mo = precip_3mo)
all_fluxes_mound_avg <- distinct(all_fluxes_mound_avg)

# merge with mound SA
mound_SA <- mound_SA[, c('sample', 'sa_m2_final', 'volume_m3_final', 'basal_area_m2', 'perimeter_m')]
mound_SA <- mound_SA[!is.na(mound_SA$sa_m2_final),]
all_fluxes_mound_avg <- merge(all_fluxes_mound_avg, mound_SA, by = "sample")

# calculate mound-level flux
all_fluxes_mound_avg$mound_CH4flux <- all_fluxes_mound_avg$mean_mound_fluxCH4 * all_fluxes_mound_avg$sa_m2_final
all_fluxes_mound_avg$mound_CO2flux <- all_fluxes_mound_avg$mean_mound_fluxCO2 * all_fluxes_mound_avg$sa_m2_final
# in units of gas per time for the whole mound structure

# EXPORT - coarse average mound fluxes
# write.csv(all_fluxes_mound_avg, "/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/finalfluxes/all_fluxes_mound_avg.csv", row.names = FALSE)

# resampled mound averages 
all_fluxes_resample_reps_avg <- all_fluxes_mound_avg[all_fluxes_mound_avg$sample %in% seasonal_list,]
# EXPORT - coarse average mound fluxes
# write.csv(all_fluxes_resample_reps_avg, "/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/finalfluxes/all_fluxes_resample_reps_avg.csv", row.names = FALSE)

```

compilation of daily dataset
```{r}
# merge CO2 and CH4 df together 
all <- full_join(CO2_daily, CH4_daily)

# convert CO2 from mmol to umol (in order to match CH4 units, conversion factor 1000umol = 1mmol)
all$flux.CO2_umol <- all$flux.CO2*1000
all <- all %>% relocate(flux.CO2_umol, .after = flux.CO2)
all <- all %>% relocate(flux.CH4, .after = flux.CO2_umol)
all <- all %>% relocate(R2.CH4, .after = R2.CO2)
all <- all %>% relocate(p.CH4, .after = p.CO2)
  
# define column for ratio of CH4:CO2
all$CH4_CO2 <-  all$flux.CH4/all$flux.CO2_umol
daily_fluxes <- all %>% relocate(CH4_CO2, .after = flux.CH4)

# rename species with scientific names 
daily_fluxes$species_s <- ifelse(daily_fluxes$species == "Ami", "A. laurensis","")
daily_fluxes$species_s <- ifelse(daily_fluxes$species == "Nasuti", "N. magnus", daily_fluxes$species_s)
daily_fluxes$species_s <- ifelse(daily_fluxes$species == "Copto", "C. acinaciformis", daily_fluxes$species_s)

# EXPORT - compiled daily fluxes
# write.csv(daily_fluxes, "/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/finalfluxes/all_daily_fluxes.csv", row.names = FALSE)
```
