==> scripts/7.flux_viz.Rmd <==
---
title: "7.flux_viz"
author: "abbey yatsko"
date: "26/08/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(ggplot2)
library(dplyr)
library(plyr)
library(ggpubr)
```

data load 
```{r}
# all fluxes
all_fluxes <- read.csv("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/finalfluxes/all_fluxes.csv")
# all fluxes - mound 
all_fluxes_mound <- read.csv("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/finalfluxes/all_fluxes_mound.csv")
# all fluxes - soil 
all_fluxes_soil <- read.csv("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/finalfluxes/all_fluxes_soil.csv")
# all fluxes - seasonal 
all_fluxes_seasonal <- read.csv("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/finalfluxes/all_fluxes_resample_reps.csv")
# all fluxes - daily 
all_fluxes_daily <- read.csv("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/finalfluxes/all_daily_fluxes.csv")
# all fluxes - mound average
all_fluxes_mound_avg <- read.csv("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/finalfluxes/all_fluxes_mound_avg.csv")
# all fluxes - naked termites 
all_fluxes_NT <- read.csv("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/finalfluxes/all_fluxes_NT.csv")
```

mound flux and species
```{r}
# overall CH4 flux from three different species
m1 <- ggplot(data = all_fluxes_mound, aes(x = species_s, y = flux.CH4, fill = species_s)) + 
  geom_boxplot()+ 
  scale_x_discrete(limits = c("N. magnus", "A. laurensis", "C. acinaciformis"))+  
  theme_classic()+
  ylim(0,28000) + # removes two outliers to make graph more readable
  ylab("CH4 flux umol/d/m2")+
  xlab("")

# overall CO2 flux from three different species
m2 <- ggplot(data = all_fluxes_mound, aes(x = species_s, y = flux.CO2, fill = species_s)) + 
  geom_boxplot()+ 
  scale_x_discrete(limits = c("N. magnus", "A. laurensis", "C. acinaciformis"))+  
  theme_classic()+
  ylab("CO2 flux mmol/d/m2")+
  xlab("")

# overall CH4:CO2 from three different species
m3 <- ggplot(data = all_fluxes_mound, aes(x = species_s, y = CH4_CO2, fill = species_s)) + 
  geom_boxplot()+ 
  scale_x_discrete(limits = c("N. magnus", "A. laurensis", "C. acinaciformis"))+  
  theme_classic()+
  ylab("CH4:CO2")+
  ylim(0,0.075)+ # removes big outlier
  xlab("")

ggarrange(m1,m2,m3, ncol = 3, common.legend = TRUE)

# table of averages by species
detach(package:plyr)
sp_avg <- all_fluxes_mound %>% group_by(species_s) %>% 
  summarize(mean_CH4flux = mean(flux.CH4), 
            mean_CO2flux = mean(flux.CO2), 
            sd_CH4flux = sd(flux.CH4), 
            sd_CO2flux = sd(flux.CO2))
  
# unit conversion umol/m2*d to ug/m2*d (final units for CH4)
# flux*(1/1000000)*(16.04)*(1000000) 
# molecular mass conversion to g, then conversion back to ug for comparison 
```

mound flux and season
```{r}
# reorder levels of campaigns 
all_fluxes_seasonal$campaign <- factor(all_fluxes_seasonal$campaign, levels = c("may22", "aug23", "nov22", "feb24"))

# select for mound data only
all_fluxes_seasonal_mound <- all_fluxes_seasonal[all_fluxes_seasonal$flux_source =="m",]

# CH4 flux 
s1 <- ggplot(data = all_fluxes_seasonal_mound, aes(x = species_s, y = flux.CH4, fill = campaign)) + 
  geom_boxplot()+ 
  scale_x_discrete(limits = c("N. magnus", "A. laurensis", "C. acinaciformis"))+  
  theme_classic()+
  ylim(0, 22186) + # removes negative outlier
  ylab("CH4 flux umol/d/m2")+
  xlab("")

# CO2 flux 
s2 <- ggplot(data = all_fluxes_seasonal_mound, aes(x = species_s, y = flux.CO2, fill = campaign)) + 
  geom_boxplot()+ 
  scale_x_discrete(limits = c("N. magnus", "A. laurensis", "C. acinaciformis"))+  
  theme_classic()+
  ylim(0,4833) + # removes negative outlier
  ylab("CO2 flux mmol/d/m2")+
  xlab("")

# overall CH4:CO2 from three different species
s3 <- ggplot(data = all_fluxes_mound, aes(x = species_s, y = CH4_CO2, fill = campaign)) + 
  geom_boxplot()+ 
  scale_x_discrete(limits = c("N. magnus", "A. laurensis", "C. acinaciformis"))+  
  theme_classic()+
  ylab("CH4:CO2")+
  ylim(0,0.06)+ # removes big outlier
  xlab("")

ggarrange(s1,s2, ncol = 2, common.legend = TRUE)
```

mound flux and temperature
```{r}
# mound flux and temperature
t1 <- ggplot(all_fluxes_seasonal_mound, aes(x = avg_respT, y = log(flux.CH4), color = campaign))+
  geom_smooth(method = 'lm', se = TRUE)+
  geom_point()+
  ylab("CH4 flux umol/d/m2")+
  xlab("Average Mound Surface Temperature (C)")+
  theme_classic()

t2 <- ggplot(all_fluxes_seasonal_mound, aes(x = avg_respT, y = flux.CO2, color = campaign))+
  geom_smooth(method = 'auto', se = TRUE)+
  ylab("CO2 flux mmol/d/m2")+
  xlab("Average Mound Surface Temperature (C)")+
  theme_classic()

ggarrange(t1, t2, ncol = 1, common.legend = TRUE)

# mound flux and time of day with temperature included
t <- all_fluxes_seasonal_mound[all_fluxes_seasonal_mound$campaign %in% c("nov22", "aug23", "feb24"),]
t$flux_start <- strptime(t$flux_start, "%H:%M")

tt <- all_fluxes_seasonal_mound[all_fluxes_seasonal_mound$campaign %in% c("may22"),]
start <- nchar(tt$flux_start) - 7
tt$flux_start_test <- substring(tt$flux_start, start, nchar(tt$flux_start))
tt$flux_start <- strptime(tt$flux_start_test, "%H:%M")
tt <- tt[, -which(names(tt) == "flux_start_test")]

all_fluxes_seasonal_mound_withtime <- rbind(t, tt)

# temperature throughout the day by season
t3 <- ggplot(all_fluxes_seasonal_mound_withtime, aes(x = as.POSIXct(flux_start), y = avg_respT, color = campaign))+
  geom_jitter()+
  geom_smooth(method = 'lm', se = TRUE)+
  ylab("Average Mound Surface Temperature (C)")+
  xlab("Time of day")+
  theme_classic()

# temperature response of gas flux by season
t4 <- ggplot(all_fluxes_seasonal_mound_withtime, aes(x = avg_respT, y = flux.CH4, color = campaign))+
  geom_point()+
  geom_smooth(method = 'lm', se = FALSE)+
  ylab("CH4 flux umol/d/m2")+
  xlab("Average Mound Surface Temperature (C)")+
  theme_classic()+
  ylim(0,25000) # address negative outlier 

ggarrange(t3, t4, ncol = 1, common.legend = TRUE)

# table of averages by season
detach(package:plyr)
season_avg <- all_fluxes_seasonal_mound %>% group_by(species_s, campaign) %>% 
  summarize(mean_CH4flux = mean(flux.CH4), 
            mean_CO2flux = mean(flux.CO2), 
            sd_CH4flux = sd(flux.CH4), 
            sd_CO2flux = sd(flux.CO2))
```

mound flux and position
```{r}
# position in mound
all_fluxes_mound$position <- ifelse(all_fluxes_mound$position == "top ", "top", all_fluxes_mound$position)
ggplot(data = all_fluxes_mound, aes(x = species_s, y = flux.CH4, fill = position)) + 
  geom_boxplot()+ 
  theme_classic()+
  ylab("CH4 flux umol/d/m2")+
  ylim(0,24000)+ # removes big outlier
  xlab("")

# direction on mound
ggplot(data = all_fluxes_mound, aes(x = directon, y = flux.CH4, fill = directon)) + 
  geom_boxplot()+ 
  theme_classic()+
  ylab("CH4 flux umol/d/m2")+
  ylim(0,24000)+ # removes big outlier
  xlab("Direction")+
  scale_x_discrete(limits = c("E", "N", "S", "W"), labels = c("E", "N", "S", "W"))
```

mound flux and size
```{r}
# reorder levels of campaigns 
all_fluxes_mound_avg$campaign <- factor(all_fluxes_mound_avg$campaign, levels = c("may22", "aug23", "nov22", "feb24"))
# graph methane flux by mound size (surface/basal area interchangeable: SA_m2/basal_area_m2)
a1 <- ggplot(all_fluxes_mound_avg, aes(x = SA_m2, y = mean_mound_fluxCH4)) +
  geom_point()+ 
  geom_smooth(method = 'lm')+
  facet_wrap(species_s~campaign, scales = "free")+ 
  theme_classic()+
  ylab("Mean mound CH4 flux (umol/d/mound)") +
  xlab("Mound surface area (m2)")

# graph carbon dioxide flux by mound size 
a2 <- ggplot(data = all_fluxes_mound_avg, aes(x = SA_m2, y = mean_mound_fluxCO2)) +
  geom_point()+ 
  geom_smooth(method = 'lm')+
  facet_grid(species_s~campaign, scales = "free_x")+ 
  theme_classic()+
  ylab("Total mound CO2 flux (mmol/d/mound)") +
  xlab("Mound surface area (m2)")

# graph CH4:CO2 flux by mound size 
a3 <- ggplot(data = all_fluxes_mound_avg, aes(x = basal_area_m2, y = mean_mound_fluxCH4_CO2)) +
  geom_point()+ 
  geom_smooth(method = 'lm')+
  scale_color_brewer("Blues")+
  facet_wrap(~species_s, scales = "free_x")+ 
  theme_classic()+
  ylab("Total mound CH4:CO2") +
  xlab("Mound basal area (m2)")

ggarrange(a1, a2, a3, ncol = 1)

# relationship between surface area and basal area (by species)
ggplot(data = all_fluxes_mound_avg, aes(x = SA_m2, y = basal_area_m2)) +
  geom_point()+ 
  geom_smooth(method = 'lm')+
  scale_color_brewer("Blues")+
  facet_wrap(~species_s, scales = "free_x")+ 
  theme_classic()+
  ylab("Mound basal area (m2)") +
  xlab("Mound surface area (m2)")
```

mound flux and wall thickness 
```{r}
df <- all_fluxes_mound[!is.na(all_fluxes_mound$wall_thickness_mm), ]

ggplot(df, aes(x = wall_thickness_mm, y = flux.CH4, color = species_s)) +
  geom_point()+ 
  theme_classic()+
  ylab("CH4 flux umol/d/m2")+
  xlab("Mound wall thickness (mm)")
```

mound flux throughout the day
```{r}
# reorder levels of time of day
all_fluxes_daily$time_of_day <- factor(all_fluxes_daily$time_of_day, levels = c("morning", "midday", "afternoon", "night"))

ggplot(all_fluxes_daily, aes(x = time_of_day, y = flux.CH4, fill = time_of_day))+
  geom_boxplot()+
  geom_point()+
  theme_classic()+
  theme(legend.position="none")+
  xlab("")+
  ylab("CH4 flux (umol/d/mound)")

# with species split out
d1 <- ggplot(all_fluxes_daily, aes(x = time_of_day, y = flux.CH4, fill = species_s))+
  geom_boxplot()+
  theme_classic()+
  xlab("")+
  ylab("CH4 flux (umol/d/mound)")

d2 <- ggplot(all_fluxes_daily, aes(x = species_s, y = flux.CH4, fill = time_of_day))+
  geom_boxplot()+
  theme_classic()+
  xlab("")+
  ylab("CH4 flux (umol/d/mound)")+
  ylim(0,25000) # removes outlier

ggarrange(d1,d2)

# temperature throughout daily cycle 
all_fluxes_daily$flux_start_test <- strptime(all_fluxes_daily$flux_start, "%H:%M")

ggplot(all_fluxes_daily, aes(x = as.POSIXct(flux_start_test), y = avg_respT))+
  geom_point()+
  theme_classic()+
  theme(legend.position="none")+ 
  xlab("Time of Day")+
  ylab("Temperature (C)")
```

naked termites 
```{r}
# termite emission factor
ggplot(all_fluxes_NT, aes(x = species, y = emission_factor, fill = species))+
  geom_boxplot()+
  geom_jitter()+
  theme_classic()+
  theme(legend.position = "none")

# flux by material 
n <- ggplot(all_fluxes_NT, aes(x = species, y = flux.CH4, fill = material))+
  geom_boxplot()+
  theme_classic()

o <- ggplot(all_fluxes_NT, aes(x = species, y = flux.CO2, fill = material))+
  geom_boxplot()+
  theme_classic()
ggarrange(n, o, ncol = 1 )
```
