==> scripts/8.flux_viz.Rmd <==
---
title: "7.flux_viz"
author: "abbey yatsko"
date: "26/08/2022"
output: html_document
---

```{r setup, include=FALSE}
library(ggplot2)
library(dplyr)
library(plyr)
library(ggpubr)
library(plotly)
library(tidyr)
library(easystats)
library(ggeffects)
library(emmeans)
library(lme4)
library(lmerTest)
library(splines)
```

data load 
```{r}
# all_fluxes_NT <- read.csv("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/finalfluxes/all_fluxes_NT.csv")
p_fluxes_f <- read.csv("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/finalfluxes/p_fluxes_f.csv")
all_fluxes_mound_avg <- read.csv("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/finalfluxes/all_fluxes_mound_avg.csv")
all_fluxes_mound <- read.csv("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/finalfluxes/all_fluxes_mound.csv")
all_fluxes_seasonal_avg <- read.csv("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/finalfluxes/all_fluxes_resample_reps_avg.csv")

# supplement (?)
all_fluxes_daily <- read.csv("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/finalfluxes/all_daily_fluxes.csv")
```

Comparison of termite fluxes at the individual (TEF) and mound level
```{r}
# Q1: How does TEF vary across study species (fluxes from individuals)
a <- ggplot(p_fluxes_f, aes(x = species, y = TEF_ug_g_h, fill = species))+
  geom_boxplot(outlier.shape = NA)+
  #geom_jitter()+
  ylab("Termite EF (ug CH4/hr/g termite)")+
  xlab("")+
  scale_x_discrete(labels = c("A. laurensis", "C. acinaciformis", "N. magnus"))+
  scale_fill_brewer(palette = "Dark2")+
  theme_classic()+
  ggtitle("a)") + 
  theme(legend.position = "none")

# Q2: How do mound fluxes vary across study species (mean fluxes from the mound)
b <- ggplot(all_fluxes_mound_avg, aes(x = species_s, y = mean_mound_fluxCH4*0.6683, fill = species_s)) + #  multiuple µmol CH₄/day by 0.6683
  geom_boxplot()+ 
  #geom_jitter()+
  scale_x_discrete(labels = c("A. laurensis", "C. acinaciformis", "N. magnus"))+
  theme_classic()+
  ylab("Mean mound CH4 flux (ug CH4/hr/m2)")+
  xlab("")+
  ggtitle("b)") + 
  scale_fill_brewer(palette = "Dark2")+
  theme(legend.position = "none")

ggarrange(a, b)

```

Mound traits influencing CH4 flux? 
```{r}
# Q1: What is the relationship between mound wall thickness and CH4 flux? 
wt <- all_fluxes_mound[!is.na(all_fluxes_mound$wall_thickness_mm), ]
wt$sample <- as.factor(wt$sample)

# flux and wall thickness
a <- ggplot(wt, aes(x = wall_thickness_mm, y = flux.CH4, color = species_s)) +
  geom_point()+ 
  geom_smooth(data = subset(wt, species_s == "C. acinaciformis"), method = "lm", se = FALSE) + 
 # facet_wrap(~species_s, scales = "free")+ 
  theme_classic()+
  ylab("CH4 flux umol/d/m2")+
  xlab("Mound wall thickness (mm)")+
  scale_color_brewer(palette = "Dark2")+
  ggtitle("a)") +
  theme(legend.position = "bottom")

# species and wall thickness 
ggplot(wt, aes(x = species_s, y = wall_thickness_mm, fill = species_s))+ 
  geom_boxplot()+
  theme_classic()+
  xlab("")+
  ylab("Mound wall thickness (mm)")+
  scale_fill_brewer(palette = "Dark2")+
  theme(legend.position = "none")

# Q2: What is the relationship between mound size and CH4 flux? 
# whole mound methane flux mound size (basal_area_m2, sa_m2_final, volume_m3_final)
b <- ggplot(all_fluxes_mound_avg, aes(x = volume_m3_final, y = mound_CH4flux, color = species_s)) + 
  geom_point(size = 4)+ 
  geom_smooth(data = subset(all_fluxes_mound_avg, species_s == "A. laurensis"), method = "lm", se = FALSE) + 
  geom_smooth(data = subset(all_fluxes_mound_avg, species_s == "N. magnus"), method = "lm", se = FALSE) + 
 # facet_wrap(~species_s, scales = "free")+ 
  scale_color_brewer(palette = "Dark2")+
  theme_classic()+
  ylab("Total mound CH4 flux (umol/d/mound)") +
  xlab("Mound volume (m3)")+
  ggtitle("b)") +
  theme(legend.position = "none")

# average mound size by species
ggplot(all_fluxes_mound_avg, aes(x = species_s, y = volume_m3_final, fill = species_s))+ # swap out basal_area_m2
  geom_boxplot()+
  theme_classic()+
  xlab("")+
  ylab("Mound volume (m3)")+
  scale_fill_brewer(palette = "Dark2")+
  theme(legend.position = "none")

ggarrange(a, b, ncol = 1)
```

How do seasons and temperature effect CH4 flux?
```{r}
# Q1: How does CH4 flux change by season?
all_fluxes_seasonal_avg$campaign <- factor(all_fluxes_seasonal_avg$campaign, levels = c("may22", "aug23", "nov22", "feb24"))
all_fluxes_seasonal_avg$species_s <- as.factor(all_fluxes_seasonal_avg$species_s)

ggplot(all_fluxes_seasonal_avg, aes(x = campaign, y = mean_mound_fluxCH4, fill = campaign))+
  geom_boxplot()+
  ylab("CH4 flux umol/d/m2")+
  xlab("")+
  scale_x_discrete(labels = c("Wet-to-dry", "Mid-dry", "Dry-to-wet", "Mid-wet"))+
  theme_classic()+
  theme(legend.position = "none")

# include species variation 
ggplot(all_fluxes_seasonal_avg, aes(x = campaign, y = mean_mound_fluxCH4, fill = species_s))+
  geom_boxplot()+
  ylab("CH4 flux umol/d/m2")+
  xlab("")+
  scale_x_discrete(labels = c("Wet-to-dry", "Mid-dry", "Dry-to-wet", "Mid-wet"))+ 
  scale_fill_manual(values = c("A. laurensis" = "#1b9e77", "C. acinaciformis" = "#d95f02", "N. magnus" =
                                 "#7570b3")) +
  theme_classic()

# Q2: How does CH4 flux change by temperature?
# LINEAR
a <- ggplot(all_fluxes_seasonal_avg, aes(x = temp, y = mean_mound_fluxCH4, color = campaign))+
  geom_point(size = 2, alpha = 0.8)+
  geom_smooth(method = 'lm', se = FALSE)+
  ylab("CH4 flux umol/d/m2")+
  xlab("Average Mound Surface Temperature (C)")+
  scale_color_brewer(palette = "Dark2")+
  ggtitle("a)") + 
  theme_classic()

# POLYNOMIAL
new_data <- expand.grid(
  temp = seq(min(all_fluxes_seasonal_avg$temp), max(all_fluxes_seasonal_avg$temp), length.out = 100),
  species_s = unique(all_fluxes_seasonal_avg$species_s),
  campaign = unique(all_fluxes_seasonal_avg$campaign)
)
                       
new_data$predicted_flux <- predict(mod_poly, newdata = new_data, re.form = NA)

b <- ggplot(new_data, aes(x = temp, y = predicted_flux, color = campaign)) +
  geom_line() + 
  #facet_wrap(~ species_s) +              # Separate plots for each species
  labs(
    x = "Temperature (°C)", 
    y = "Predicted CH4 Flux (umol/d/m2)",
    title = "Polynomial Relationship between Temperature and CH4 Flux"
  ) +
  scale_color_brewer(palette = "Dark2")+
  ggtitle("b)") + 
  theme_classic()

ggarrange(a, b, ncol = 1)
```

















Other potentially interesting things 
------------------------------------
fluxes from termites + mound material + together 
```{r}
# Q3: What are the fluxes associated with termites in the mound vs. termites outside mound
# select columns from p_fluxes_f
NT_ch4_long <- p_fluxes_f %>% select(species, flux.CH4_A, flux.CH4_T, flux.CH4_M)
NT_ch4_long <- gather(NT_ch4_long, key = "source", value = "flux", -species)
NT_ch4_long$source <- factor(NT_ch4_long$source, levels = c("flux.CH4_A", "flux.CH4_T", "flux.CH4_M"))

NT_co2_long <- p_fluxes_f %>% select(species, flux.CO2_A, flux.CO2_T, flux.CO2_M)
NT_co2_long <- gather(NT_co2_long, key = "source", value = "flux", -species)
NT_co2_long$source <- factor(NT_co2_long$source, levels = c("flux.CO2_A", "flux.CO2_T", "flux.CO2_M"))

c <- ggplot(NT_ch4_long, aes(x = species, y = flux, fill = source))+
  geom_boxplot(outlier.shape = NA)+
  ylab("CH4 flux (umol/d/m2)")+
  xlab("")+
  scale_x_discrete(labels = c("A. laurensis", "C. acinaciformis", "N. magnus"))+
  scale_fill_brewer(palette = "RdYlGn", labels = c("T+M", "T", "M"))+
  theme_classic()

d <- ggplot(NT_co2_long, aes(x = species, y = flux, fill = source))+
  geom_boxplot(outlier.shape = NA)+
  ylab("CO2 flux (mmol/d/m2)")+
  xlab("")+
  scale_x_discrete(labels = c("A. laurensis", "C. acinaciformis", "N. magnus"))+
  scale_fill_brewer(palette = "RdYlGn", labels = c("T+M", "T", "M"))+
  theme_classic()

ggarrange(c,d)
```


splines of different df 
```{r}
spline3 <- lmer(mean_mound_fluxCH4 ~ ns(temp, df = 3) + species_s + campaign + (1 | sample), data = all_fluxes_seasonal_avg)
summary(spline3)
residuals <- resid(spline3)
fitted_values <- fitted(spline3)
plot(fitted_values, residuals, main = "Residuals vs Fitted Values",
     xlab = "Fitted Values", ylab = "Residuals")
abline(h = 0, col = "red")
qqnorm(resid(spline3))
qqline(resid(spline3))

# how many df to use? seems like anything above 5 does not make much sense (lines start to get a bit weird)
spline4 <- lmer(mean_mound_fluxCH4 ~ ns(temp, df = 4) + species_s + campaign + (1 | sample), data = all_fluxes_seasonal_avg)
summary(spline4)

spline5 <- lmer(mean_mound_fluxCH4 ~ ns(temp, df = 5) + species_s + campaign + (1 | sample), data = all_fluxes_seasonal_avg)
summary(spline5)

spline6 <- lmer(mean_mound_fluxCH4 ~ ns(temp, df = 6) + species_s + campaign + (1 | sample), data = all_fluxes_seasonal_avg)
summary(spline6)
residuals <- resid(spline6)
fitted_values <- fitted(spline6)
plot(fitted_values, residuals, main = "Residuals vs Fitted Values",
     xlab = "Fitted Values", ylab = "Residuals")
abline(h = 0, col = "red")
qqnorm(resid(spline6))
qqline(resid(spline6))

spline7 <- lmer(mean_mound_fluxCH4 ~ ns(temp, df = 7) + species_s + campaign + (1 | sample), data = all_fluxes_seasonal_avg)
summary(spline7)

AIC(spline3, spline4, spline5, spline6, spline7, spline10)
```

PROPORTIONAL FLUXES - from naked termite measurements, what proportion of the total flux is from termites vs mounds? 
```{r}
df <- p_fluxes_f[c("species", "prop_T_CH4", "prop_M_CH4", "prop_T_CO2", "prop_M_CO2")]
df_long <- df %>%
  pivot_longer(cols = starts_with("prop_"), 
               names_to = c(".value", "flux"), 
               names_pattern = "prop_(.*)_(.*)") %>%
  pivot_longer(cols = c(T, M), 
               names_to = "material", 
               values_to = "prop") 

ggplot(df_long[df_long$flux == "CH4",], aes(x = species, y = prop, fill = material))+
  geom_boxplot()+
  ylab("Proportion CH4 flux")+
  xlab("")+
  scale_x_discrete(labels = c("A. laurensis", "C. acinaciformis", "N. magnus"))+
  scale_fill_brewer(palette = "Dark2")+
  theme_classic()

ggplot(df_long[df_long$flux == "CO2",], aes(x = species, y = prop, fill = material))+
  geom_boxplot()+
  ylab("Proportion CO2 flux")+
  xlab("")+
  scale_x_discrete(labels = c("A. laurensis", "C. acinaciformis", "N. magnus"))+
  scale_fill_brewer(palette = "Dark2")+
  theme_classic()

# test what the ratio of termite:mound mass is - best if consistent across the df (seems to be largely true)
p_fluxes_f$termite_mound_ratio <- p_fluxes_f$mass_termite_g / p_fluxes_f$mass_mound_material_g
ggplot(p_fluxes_f, aes(x = species, y = termite_mound_ratio, fill = species))+
  geom_boxplot()+
  ylab("Termite mass:Mound mass")+
  xlab("")+
  scale_x_discrete(labels = c("A. laurensis", "C. acinaciformis", "N. magnus"))+
  scale_fill_brewer(palette = "Dark2")+
  theme_classic()+
  theme(legend.position = "none")
```

Temperature variation by campaign
```{r}
# temperatures by campaign - this just shows seasonal variability
ggplot(all_fluxes_mound_avg, aes(x = campaign, y = temp, fill = campaign))+
  geom_boxplot()+
  geom_jitter()+
  ylab("Average Mound Surface Temperature (C)")+
  xlab("")+
  scale_fill_brewer(palette = "Dark2")+
  theme_classic()+
  theme(legend.position = "none")

```

Code for daily cycle question
```{r}
# temperature changes throughout the day
t <- all_fluxes_mound[all_fluxes_mound$campaign %in% c("nov22", "aug23", "feb24"),]
t$flux_start <- strptime(t$flux_start, "%H:%M")

tt <- all_fluxes_mound[all_fluxes_mound$campaign %in% c("may22"),]
start <- nchar(tt$flux_start) - 7
tt$flux_start_test <- substring(tt$flux_start, start, nchar(tt$flux_start))
tt$flux_start <- strptime(tt$flux_start_test, "%H:%M")
tt <- tt[, -which(names(tt) == "flux_start_test")]

all_fluxes_withtime <- rbind(t, tt)

# temperature throughout the day by season
ggplot(all_fluxes_withtime, aes(x = as.POSIXct(flux_start), y = avg_respT, color = campaign))+
  geom_jitter()+
  geom_smooth(method = 'loess', se = TRUE)+
  ylab("Average Mound Surface Temperature (C)")+
  xlab("Time of day")+
  theme_classic()

# table of averages by season (all mounds, not those with repeat measurements)
detach(package:plyr)
season_avg <- all_fluxes_mound_avg %>% group_by(species_s, campaign) %>% 
  summarize(mean_CH4flux = mean(mean_mound_fluxCH4), 
            mean_CO2flux = mean(mean_mound_fluxCO2), 
            sd_CH4flux = sd(mean_mound_fluxCH4), 
            sd_CO2flux = sd(mean_mound_fluxCO2), 
            count = n())
```

Intramound flux variation 
```{r}
# INTRAMOUND VARIATION OF MOUND FLUX
all_fluxes_mound$position <- ifelse(all_fluxes_mound$position == "top ", "top", all_fluxes_mound$position)
ggplot(data = all_fluxes_mound, aes(x = species_s, y = flux.CH4, fill = position)) + 
  geom_boxplot()+ 
  theme_classic()+
  ylab("CH4 flux umol/d/m2")+
  ylim(0,24000)+ # removes big outlier
  xlab("")

# direction on mound
ggplot(data = all_fluxes_mound, aes(x = directon, y = flux.CH4, fill = directon)) + 
  geom_boxplot()+ 
  theme_classic()+
  ylab("CH4 flux umol/d/m2")+
  ylim(0,24000)+ # removes big outlier
  xlab("Direction")+
  scale_x_discrete(limits = c("E", "N", "S", "W"), labels = c("E", "N", "S", "W"))

# TIME OF DAY AND FLUX
# reorder levels of time of day
all_fluxes_daily$time_of_day <- factor(all_fluxes_daily$time_of_day, levels = c("morning", "midday", "afternoon", "night"))

# time by flux for species
ggplot(all_fluxes_daily, aes(x = species_s, y = flux.CH4, fill = time_of_day))+
  geom_boxplot()+
  theme_classic()+
  xlab("")+
  ylab("CH4 flux (umol/d/mound)")+
  ylim(0,25000) # removes outlier

# calculate mound average for flux and temp 
all_fluxes_daily$sampleTOD <- paste(all_fluxes_daily$sample, all_fluxes_daily$time_of_day, sep = "-")
daily_avg <- all_fluxes_daily %>% group_by(sampleTOD) %>% 
  summarise(mean_mound_fluxCH4=mean(flux.CH4),
            mean_mound_fluxCO2=mean(flux.CO2_umol), 
            mean_mound_fluxCH4_CO2=mean(CH4_CO2), 
            time_of_day = time_of_day, 
            species_s = species_s,
            sample = sample,
            temp = mean(avg_respT))
daily_avg_f <- distinct(daily_avg)

# temp by flux for species at mound-level average
ggplot(daily_avg_f, aes(x = temp, y = mean_mound_fluxCH4, color = sample))+
  geom_point()+
  geom_smooth(method = "loess")+
  facet_wrap(~species_s, scales = "free_y")+
  theme_classic()+
  xlab("Temperature (C)")+
  ylab("CH4 flux (umol/d/mound)")
```
