==> scripts/9.flux_models.Rmd <==
---
title: "8.flux_models"
author: "abbey yatsko"
date: "2024-03-15"
output: html_document
---

libraries
```{r}
library(ggplot2)
library(lme4)
library(car)
library(easystats)
library(ggeffects)
library(vegan)
library(partR2)
library(patchwork)
library(emmeans)
library(lmerTest)
```

data load 
```{r}
p_fluxes_f <- read.csv("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/finalfluxes/p_fluxes_f.csv")
all_fluxes_mound_avg <- read.csv("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/finalfluxes/all_fluxes_mound_avg.csv")
all_fluxes_mound <- read.csv("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/finalfluxes/all_fluxes_mound.csv")

all_fluxes_resample_reps_avg <- read.csv("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/finalfluxes/all_fluxes_resample_reps_avg.csv")

all_fluxes_daily <- read.csv("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/finalfluxes/all_daily_fluxes.csv")
```

Determine if there is significant within mound variation 
```{r}
# random effects model 
model <- lmer(flux.CH4 ~ 1 + (1 | ID_mound), data = all_fluxes_mound)
summary(model)
ranova(model) # signifiacnt random effect means that intramound variation matters

# intraclass correlation coefficient 
icc <- as.numeric(VarCorr(model)$ID_mound) / 
       (as.numeric(VarCorr(model)$ID_mound) + attr(VarCorr(model), "sc")^2)

# interpretation: ~34.8% of the total variation in CH4 flux is due to differences between mounds, while ~65.2% is due to variation within mounds

# need to retain the random effect in subsequent models

# visualisation of within and between mound variation 
ggplot(all_fluxes_mound, aes(x = ID_mound, y = flux.CH4)) +
  geom_boxplot() +
  theme_minimal() +
  labs(title = "CH4 Flux Variation Within and Between Mounds",
       x = "Mound ID",
       y = "CH4 Flux") +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))
```

Comparison of termite fluxes at the individual and mound level
```{r}
# Q1: How does TEF vary across study species (fluxes from individuals)
# test for differences between species 
p_fluxes_f$species <- as.factor(p_fluxes_f$species)
mod <- aov(TEF_ug_g_h ~ species, data = p_fluxes_f)
summary(mod)

# check assumptions 
qqnorm(resid(mod))
qqline(resid(mod))
plot(fitted(mod), resid(mod))

# determine which species are different
emmeans(mod, pairwise ~ species, adjust = "tukey")

# Q2: How do mound fluxes vary across study species
all_fluxes_mound$ID_mound <- as.factor(all_fluxes_mound$ID_mound)
all_fluxes_mound$species_s <- as.factor(all_fluxes_mound$species_s)

mod <- lmer(flux.CH4 ~ species_s + (1 | ID_mound), data = all_fluxes_mound)
summary(mod)
Anova(mod)
emmeans(mod, pairwise ~ species_s)

# check assumptions 
qqnorm(resid(mod))
qqline(resid(mod))
plot(fitted(mod), resid(mod))

# QUESTION: to include negative/zero fluxes in the model? response might need transformation... 
```

How do mound traits contribute to CH4 flux? 
```{r}
# Q1: What is the relationship between mound wall thickness and CH4 flux? 
# flux influenced by wall thickness
wt <- all_fluxes_mound[!is.na(all_fluxes_mound$wall_thickness_mm), ]
wt$sample <- as.factor(wt$sample)

# average wall thickness by species significantly differs
mod <- aov(wall_thickness_mm ~ species_s, data = wt)
summary(mod)
emmeans(mod, pairwise ~ species_s)

# centre wall_thickness to account for species differences in mound wall thickness
wt$wall_thickness_centered <- ave(wt$wall_thickness_mm, wt$species_s, FUN = scale)

# consider only positive fluxes in models 
wt_pos <- wt[wt$flux.CH4 > 0, ]

ggplot(wt_pos, aes(x = wall_thickness_centered, y = log(flux.CH4), color = species)) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_classic() +
  scale_color_brewer(palette = "Dark2")+
  labs(x = "Wall Thickness (mean centred)",
       y = "CH4 Flux")

mod <-  lmer(log(flux.CH4) ~ wall_thickness_centered*species_s + (1 | ID_mound), data = wt_pos)
summary(mod)
Anova(mod)

# check assumptions 
qqnorm(resid(mod))
qqline(resid(mod))
plot(fitted(mod), resid(mod))

# posthoc emmeans 
emms <- emmeans(mod, ~ wall_thickness_centered | species_s)
summary(emms)
emtr <- emtrends(mod, ~ species_s, var = "wall_thickness_centered")
summary(emtr)


# Q2: What is the relationship between mound size and CH4 flux? 
# average mound size by species
mod <- aov(volume_m3_final ~ species_s, data = all_fluxes_mound_avg)
summary(mod)
emmeans(mod, pairwise ~ species_s)

all_fluxes_mound_avg$volume_m3_final_centered <- ave(all_fluxes_mound_avg$volume_m3_final, all_fluxes_mound_avg$species_s, FUN = scale)

ggplot(all_fluxes_mound_avg, aes(x = volume_m3_final_centered, y = (mound_CH4flux), color = species_s)) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_classic() +
  scale_color_brewer(palette = "Dark2")+
  labs(x = "Mound Volume (mean centred)",
       y = "CH4 Flux")

mod <- lmer(mound_CH4flux ~ volume_m3_final_centered*species_s + (1 | sample), data = all_fluxes_mound_avg)
summary(mod)
Anova(mod)

qqnorm(resid(mod))
qqline(resid(mod))
plot(fitted(mod), resid(mod))

emms <- emmeans(mod, ~ volume_m3_final_centegured | species_s)
summary(emms)
```

How do seasons and temperature effect CH4 flux?
```{r}
# How does CH4 flux change by season and temperature?
# LINEAR
mod <- lmer(mean_mound_fluxCH4 ~ temp + campaign + species_s + (1 | sample) + (1 | species_s), data = all_fluxes_resample_reps_avg)
summary(mod)
Anova(mod)

tab_model(mod)

emm <- emmeans(mod, ~ campaign)
pairs(emm)

# mean mean_mound_fluxCH4 by campaign table
mean_mound_fluxCH4_by_campaign <- all_fluxes_resample_reps_avg %>%
  group_by(campaign) %>%
  summarise(mean_mound_fluxCH4 = mean(mean_mound_fluxCH4, na.rm = TRUE))

# POLYNOMIAL 
mod_poly <- lmer(mean_mound_fluxCH4 ~ poly(temp, 2) + campaign + (1 | sample) + (1 | species_s), data = all_fluxes_resample_reps_avg)
summary(mod_poly)
Anova(mod_poly, type = "III")

anova(mod, mod_poly)

emmeans(mod_poly, pairwise ~ campaign)
```

######
EXTRA 
######

```{r}
# relationship between relabund and wall thickness 
ggplot(flux_relabund_pos, aes(x = wall_thickness_mm, y = resampled_relabund, color = species_s))+
  geom_point()+ 
  geom_smooth(method = "lm") + 
  theme_classic()

mod <- lm(resampled_relabund ~ wall_thickness_mm*species_s, data = flux_relabund_pos)
summary(mod)
# average mound relabund and size? 

# Copto: relationhsip between wall thickness and volume 
ggplot(all_fluxes_mound_avg, aes(x = wall_thickness_mm, y = volume_m3_final))+
  geom_point()+
  facet_wrap(~species_s, scales = "free")+
  geom_smooth(method = "lm")+
  theme_minimal()

p_fluxes_f %>%
  group_by(species) %>%
  summarise(TEF = mean(TEF_ug_g_h)) -> TEF

all_fluxes_mound_avg %>%
  group_by(species_s) %>%
  summarise(size = mean(volume_m3_final), 
            min = min(volume_m3_final), 
            max = max(volume_m3_final), 
            flux = mean(mound_CH4flux)) -> size

```

```{r}
# CH4 flux from mounds if you just used TEF * biomass 
p_fluxes_f$t_biomass_g_m3 <- p_fluxes_f$mass_termite_g / p_fluxes_f$material_volume_cm3 * 10e6

# calculate average t_biomass_g_m3 by species
t_biomass <- p_fluxes_f %>% 
  group_by(species) %>% 
  summarise(t_biomass_g_m3 = mean(t_biomass_g_m3), 
            TEF_ug_g_h = mean(TEF_ug_g_h))

t_biomass$species_s <- c("A. laurensis", "C. acinaciformis", "N. magnus")
t_biomass <- t_biomass[, -1]

all_fluxes_mound_avg <- left_join(all_fluxes_mound_avg, t_biomass, by = "species_s")

all_fluxes_mound_avg$mound_CH4flux_TEF <- all_fluxes_mound_avg$t_biomass_g_m3 *all_fluxes_mound_avg$volume_m3_final * all_fluxes_mound_avg$TEF_ug_g_h 

# data to long format for the columns mound_CH4flux_TEF and mound_CH4flux
all_fluxes_mound_avg_long <- all_fluxes_mound_avg %>% 
  pivot_longer(cols = c(mound_CH4flux_TEF, mound_CH4flux), names_to = "flux_type", values_to = "flux")

# print each comparison figure to pdf
pdf("mound_flux_plots.pdf", onefile = TRUE)  # onefile = TRUE ensures multiple plots in one PDF

# Loop through each unique mound ID
unique_mounds <- unique(all_fluxes_mound_avg_long$ID_mound)

for(mound in unique_mounds) {
  # Subset the data for the current mound
  mound_data <- subset(all_fluxes_mound_avg_long, ID_mound == mound)
  
  # Create the plot for the current mound
  plot <- ggplot(mound_data, aes(x = flux_type, y = flux, color = flux_type)) +
    geom_point(alpha = 0.7, size = 3) +
    theme_minimal() +
    labs(
      x = "Species",
      y = "CH4 Flux (ug/m2/h)",
      fill = "Flux Type",
      title = paste("Mound ID:", mound)  # Add a title to each plot
    )
  
  # Print the plot to the PDF
  print(plot)
}

# Close the PDF device
dev.off()
```

```{r}
# create new dataframe for all_fluxes_mound_avg$species_s == "A.laurensis"
z <- wt[wt$species_s == "N. magnus", ]

# average mound level flux by species 
species_flux_size <- all_fluxes_mound_avg %>% 
  group_by(species_s) %>% 
  summarise(mean_mound_CH4flux = mean(mound_CH4flux), 
            mean_volume_m3 = mean(volume_m3_final), 
            n = n_distinct(sample))


# Given values in mg CH4 mound⁻¹ hr⁻¹
values_mg <- c(0.07, 2.7, 0.023, 0.08, 0.069, 10.3)


# Molecular weight of CH4 (in g/mol)
molecular_weight_CH4 <- 16.04

# Convert from mg CH4 to µmol CH4 hr⁻¹ mound⁻¹
values_umol <-  (values_mg * 1000 / molecular_weight_CH4) 
values_umol


```
