==> scripts/11.metagenome_analysis.Rmd <==
---
title: "meth_analysis"
author: "abbey yatsko"
date: "2024-11-18"
output: html_document
---

workspace
```{r}
library(ggplot2)
library(dplyr)
library(tidyr)
library(stringr)
library(ggpubr)
library(datawizard)
library(lme4)
library(car)
library(lmerTest)

metadata <- read.csv("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/data_out/metagenomes_metadata.csv")
top_e <- read.csv("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/data_out/q1_top_emitters.csv")
var_e <- read.csv("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/data_out/q2_variable_emitters.csv")
fluxes <- read.csv("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/finalfluxes/all_fluxes_mound.csv")
targets <- read.csv("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/metagenomes/targets.csv")
wt <- fluxes[!is.na(fluxes$wall_thickness_mm), ]
ko <- read.delim("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/metagenomes/ko.table(ReadsPerSample).txt", header = TRUE, sep = "\t")

# pathway <- read.delim("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/metagenomes/pathway.table(ReadsPerSample).txt", header = TRUE, sep = "\t")
# cazymes <- read.delim("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/metagenomes/CAZy.single.table(ReadsPerSample).txt", header = TRUE, sep = "\t")

resampled_data <- read.csv("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/metagenomes/ko_resampled_data.csv")
resampled_data <- resampled_data[, -1]
```

data clean 
```{r}
fluxes <- fluxes[fluxes$campaign == "may22", ]
fluxes$extract_ID <- paste(fluxes$sample,"-",fluxes$tag, "-", fluxes$position, sep = "")
wt_merge <- wt[, c("ID_measurement", "wall_thickness_mm")]
x <- fluxes[, c("ID_measurement", "extract_ID", "flux.CH4", "avg_respT")]
x <- left_join(x, wt_merge, by = "ID_measurement")

# merge x and metadata 
metadata <- left_join(metadata, x, by = "extract_ID")
```

distribution of total reads per sample - ko table
```{r}
# ko number
integer_cols <- sapply(ko, is.integer)
total_reads <- colSums(ko[, integer_cols])

total_reads_ko <- data.frame(
  Sample = names(total_reads),
  Sum = total_reads,
  row.names = NULL
)

# plot distribution of total reads per sample
ggplot(total_reads_ko, aes(x = reorder(Sample, -Sum), y = Sum)) +
  geom_bar(stat = "identity") +
  labs(title = "KO table - reads per sample", x = "Total reads", y = "Frequency") +
  theme_classic() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) 
```

identify standardisation value lower limit total reads (ko)
```{r}
ko_ok <- ko[ , !names(ko) %in% c("A108","A109", "A107", "A38", "A40")]
ko_ok <- ko_ok[, -1]

ko_ok_long <- ko_ok %>%
  pivot_longer(cols = -Annotation, names_to = "sequencing_id", values_to = "reads")

# # resample for lower limit  
total_reads_ko <- total_reads_ko[!(total_reads_ko$Sample %in% c("A108","A109", "A107", "A38", "A40")), ]
lower_limit <- min(total_reads_ko$Sum)

# NOTE run resampling on big computadora
# set.seed(123)
# resampled_data <- ko_ok_long %>%
#   group_by(sequencing_id) %>%
#   mutate(prob = reads / sum(reads)) %>%  # Calculate probabilities for sampling
#   slice_sample(n = lower_limit, replace = TRUE, weight_by = prob) %>%  # Sample rows
#   count(sequencing_id, Annotation, name = "resampled_reads") %>%  # Summarize the results
#   ungroup()

# sum of reads by sample (should equal lower_limit for all samples)
# total_reads_resampled <- resampled_data %>% 
#   group_by(sequencing_id) %>%
#   summarise(Sum = sum(resampled_reads))

# write.csv(resampled_data, "\\\\am_dsm1821\\BAMspace\\6_staff_folders\\Abbey\\metagenomics\\metagenomescopy\\ko_resampled_data.csv")
# resampled_data <- read.csv("/Users/abbeyyatsko/Desktop/repos/TMI_flux/data/metagenomes/ko_resampled_data.csv")
#resampled_data <- resampled_data[, -1]
```

compare relative abundance for ko (non-standardised v standardised)
```{r}
ko_ok_long <- left_join(ko_ok_long, total_reads_ko, by = c("sequencing_id" = "Sample"))
ko_ok_long$relabund <- ko_ok_long$reads/ko_ok_long$Sum
ko_ok_long %>% group_by(sequencing_id) %>% summarise(Sum_prop = sum(relabund)) -> out

# join with resampled data 
ko_ok_long <- left_join(ko_ok_long, resampled_data, by = c("sequencing_id", "Annotation"))
ko_ok_long$resampled_sum <- lower_limit
ko_ok_long$resampled_relabund <- ko_ok_long$resampled_reads/ko_ok_long$resampled_sum

# bar plot summary of relative abundance
a <- ggplot(ko_ok_long, aes(x = sequencing_id, y = relabund, fill = Annotation)) +
  geom_bar(stat = "identity") +
  labs(title = "Pathway table - relative abundance per sample", x = "Sample", y = "Relative abundance") +
  theme_classic() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1), legend.position = "none")

b <- ggplot(ko_ok_long, aes(x = sequencing_id, y = resampled_relabund, fill = Annotation)) +
  geom_bar(stat = "identity") +
  labs(title = "Pathway table - relative abundance per sample (resampled)", x = "Sample", y = "Relative abundance") +
  theme_classic() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1), legend.position = "none")

ggarrange(a, b, ncol = 2)

pmoa <- ko_ok_long[grepl("pmoA", ko_ok_long$Annotation), ]

c <- ggplot(pmoa, aes(x = sequencing_id, y = relabund, fill = sequencing_id)) +
  geom_bar(stat = "identity") +
  labs(title = "Relative abundance of methane pathway per sample", x = "Sample", y = "Relative abundance") +
  theme_classic() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1), legend.position = "none")

d <- ggplot(pmoa, aes(x = sequencing_id, y = resampled_relabund, fill = sequencing_id)) +
  geom_bar(stat = "identity") +
  labs(title = "Relative abundance of methane pathway per sample (resampled)", x = "Sample", y = "Relative abundance") +
  theme_classic() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1), legend.position = "none")

ggarrange(c, d, ncol = 2)

# direct comparison with data/resampled data
pmoa$relabund_diff <- pmoa$relabund - pmoa$resampled_relabund
ggplot(pmoa, aes(x = sequencing_id, y = relabund_diff, fill = sequencing_id)) +
  geom_bar(stat = "identity") +
  labs(title = "Difference in relative abundance from resampling", x = "Sample", y = "Relative abundance") +
  theme_classic() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1), legend.position = "none")

mean_relabund <- mean(pmoa$resampled_relabund, na.rm = TRUE)
```

visualise relative abundance of all ko numbers for context
```{r}
# average relative abundance by pathway 
ko_ok_long %>% group_by(Annotation) %>% 
  summarise(mean_relabund = mean(relabund, na.rm = TRUE), 
            mean_resampled_relabund = mean(resampled_relabund, na.rm = TRUE)) -> each_pathway_relabund

ggplot(each_pathway_relabund, aes(x = Annotation, y = mean_resampled_relabund, fill = Annotation)) +
  geom_bar(stat = "identity") +
  labs(title = "Mean relative abundance per pathway", x = "Pathway", y = "Relative abundance") +
  theme_classic() +
  theme(axis.text.x = element_blank(),  # Remove x-axis labels
        axis.ticks.x = element_blank(), # Optionally remove x-axis ticks
        legend.position = "none") + 
  geom_hline(yintercept = mean_relabund, linetype = "dashed") + # mean methane metabolism relative abundance
  ylim(0, 0.001) # limits y axis to show comparison of pmoa

ggplot(each_pathway_relabund, aes(x = reorder(Annotation, -mean_resampled_relabund), y = mean_resampled_relabund)) +
  geom_bar(stat = "identity") +
  labs(title = "KO table - reads per sample", x = "Total reads", y = "Frequency") +
  theme_classic() +
  theme(axis.text.x = element_blank(),  # Remove x-axis labels
        axis.ticks.x = element_blank(), # Optionally remove x-axis ticks
        legend.position = "none") 
```

merge metadata with methane metabolism pathway relative abundances, flux ~ relabund
```{r}
# join pmoa relabund with flux data
flux_relabund <- left_join(pmoa, metadata, by = c("sequencing_id"))
flux_relabund$species_s <- as.factor(flux_relabund$species_s)

# remove samples with no methane flux data
flux_relabund <- flux_relabund[!is.na(flux_relabund$flux.CH4), ]

# standardise data by rescaling with mean = 0 and sd = 1 
flux_relabund <- flux_relabund %>%
  mutate(resampled_relabund_scaled = standardize(relabund))

flux_relabund$resampled_relabund_scaled <- as.numeric(flux_relabund$resampled_relabund_scaled)

# write.csv(flux_relabund, "/Users/abbeyyatsko/Downloads/flux_relabund.csv", row.names = FALSE)

# consider only positive values in the model 
flux_relabund_pos <- flux_relabund[flux_relabund$flux.CH4 > 0, ]

# does greater relative abundance of methane metabolism pathway correlate with lower methane flux?
ggplot(flux_relabund, aes(x = resampled_relabund_scaled, y = (flux.CH4))) + 
  geom_point() + 
  geom_smooth(method = "lm") + 
  theme_classic() + 
  xlab("Relative abundance of pmoA (mean centred)") + 
  ylab("CH4 Flux")

mod <- lmer(log(flux.CH4) ~ resampled_relabund_scaled + species_s + (1|sample), data = flux_relabund_pos)
summary(mod)
Anova(mod)

# check assumptions 
qqnorm(resid(mod))
qqline(resid(mod))
plot(fitted(mod), resid(mod))

random_effects <- ranef(mod)$sample[, 1]
qqnorm(random_effects)
qqline(random_effects)

# species difference in relative abundance
ggplot(flux_relabund_pos, aes(x = species_s, y = relabund, fill = species_s)) +
  geom_boxplot() +
  theme_classic() +
  scale_fill_brewer(palette = "Dark2")+
  xlab("") + 
  ylab("Relative abundance of pmoA") + 
  theme(legend.title = element_blank())

mod <- aov(relabund ~ species_s, data = flux_relabund_pos)
summary(mod)
emmeans(mod, pairwise ~ species_s)
```

Q1: are there differences in relative abundance of methane pathway between low/high flux mounds within species
```{r}
data_q1 <- flux_relabund_pos %>%
  filter(!is.na(rank))
data_q1$rank <- as.factor(data_q1$rank)

# variability of pathway expression and CH4 flux within each sample 
ggplot(data_q1, aes(x = sample, y = resampled_relabund, colour = flux.CH4)) +
  geom_point(size = 5, alpha = .5) + 
  theme_classic() + 
  facet_wrap(~ rank) + 
  scale_color_gradientn(colors = rainbow(5))+
  theme(axis.text.x = element_text(angle = 45, hjust = 1), legend.position = "none")

# species variation in CH4 flux and pathway expression for low and high emitters and by species 
a <- ggplot(data_q1, aes(x = species, y = flux.CH4, fill = rank)) +
  geom_boxplot() +
  theme_classic()
  
b <- ggplot(data_q1, aes(x = species, y = resampled_relabund, fill = rank)) +
  geom_boxplot() +
  theme_classic()

ggarrange(a, b)

# for each species, are there differences in relabund of methane pathway between high and low flux mounds? 
mod <- lmer(log(relabund) ~ rank + species + (1 | sample), data = data_q1)
summary(mod)
Anova(mod)

ggplot(data_q1, aes(x = rank, y = relabund, fill = rank)) +
  geom_boxplot() +
  theme_classic()

# check assumptions 
qqnorm(resid(mod))
qqline(resid(mod))
plot(fitted(mod), resid(mod))

random_effects <- ranef(mod)$sample[, 1]
qqnorm(random_effects)
qqline(random_effects)
```

Q2: do low and high emitters have different pathway expression
```{r}
data_q2 <- flux_relabund_pos %>%
  filter(!is.na(variable_emitters))

# summarize counts per sample in a table 
data_q2 %>% group_by(sample) %>% 
  summarise(count = n()) 

# for each sample, calculate max(relabund) - min(relabund)
data_q2 %>% group_by(sample, species) %>% 
  summarise(diff_relabund = max(resampled_relabund) - min(resampled_relabund), 
            range = range) -> diff_relabund

x <- unique(diff_relabund)

# relationship between range of CH4 flux and range of relabund 
ggplot(x, aes(x = range, y = diff_relabund)) + 
  geom_point() + 
  geom_smooth(method = "lm") + 
  theme_classic() + 
  xlab("Range of CH4 flux") +
  ylab("Range of relabund")

mod <- lm(log(diff_relabund) ~ range, data = x)
summary(mod)
# doesn't really matter? 
```








#################
EXTRAS
#################

models with additional predictors
```{r}
# add wall thickness and temperature to the model
mod_without_wall <- lmer(flux.CH4 ~ resampled_relabund_scaled + species + avg_respT + (1|sample), data = flux_relabund)
Anova(mod_without_wall, type = "III")
ggpredict(mod_without_wall, terms = c("species", "avg_respT")) %>% plot(add.data = TRUE)
VarCorr(mod_without_wall)

mod_with_wall <- lmer(flux.CH4 ~ resampled_relabund_scaled + species + avg_respT + wall_thickness_mm + (1|sample), 
                      data = flux_relabund[!is.na(flux_relabund$wall_thickness_mm), ])
Anova(mod_with_wall, type = "III")



```
